### 2020-12-08 #2898 @whedon
[PRE REVIEW]: NLMech: Implementation of finite difference/meshfreediscretization of nonlocal fracture models
**Submitting author:** @diehlpk (<a href="http://orcid.org/0000-0003-3922-8419">Patrick Diehl</a>)
**Repository:** <a href="https://github.com/nonlocalmodels/NLMech" target ="_blank">https://github.com/nonlocalmodels/NLMech</a>
**Version:** v0.1.0-rc1
