### 2020-12-14 #2898 @ilyasst
I am interested in the subject of this publication and can review it in the following weeks. However, I'd like to check with you @Kevin-Mattheus-Moerman that there is no conflict of interest here as I know @diehlpk quite well. We have worked together in the past (same lab for some time) and are currently collaborating on some projects.
