### 2020-12-15 #2898 @Kevin-Mattheus-Moerman
@ctdegroot @chennachaos @tjhei, @likask would you be interested in reviewing this work for JOSS? It features nonlocal fracture models and finite difference/meshfree methods.

FYI we can be flexible the coming month regarding timing (since it is a holiday season for many), so if you are able to accept this invitation but start review at a later time, or take quite a bit more time than usual, that is fine too. Thanks!

I realize I've invited some of you for multiple submissions. We are really thankful for your review efforts. Please do let me know if you feel you are receiving too many invites from us.
