### 2020-12-16 e1f214a
Fix the wrong output number after restart
