### 2020-12-17 #2898 @tjhei
> @tjhei would you be interested in reviewing this work for JOSS?

The next 2-3 months are quite inconvenient for me. Sorry.
