### 2020-12-18 ca46c9f
Add a load over line which is scaled by the distance
