### 2020-12-18 74fcdac
Loading (#112)
* Add a load over line which is scaled by the distance

* Error handling

* Clean up

* Format
