### 2020-12-31 #2898 @Kevin-Mattheus-Moerman
@ctdegroot, @likask, @jritta would you be interested in reviewing this work for [JOSS](https://joss.theoj.org)? It features nonlocal fracture models and finite difference/meshfree methods.

The review focuses on [the software](https://github.com/nonlocalmodels/NLMech) and this [short paper](https://github.com/openjournals/joss-papers/blob/joss.02898/joss.02898/10.21105.joss.02898.pdf).

This is a pre-review issue in which we find and assign editors/reviewers. Once sufficient reviewers are recruited a dedicated review issue is opened.

Let me know if you are interested. Thanks, and best wishes for 2021!
