### 2021-02-01 #2898 @ctdegroot
Normally I would say yes, but I have another JOSS review in progress at this time, so I can't commit to finishing this in a timely manner. Best of luck!
