### 2021-02-04 ee63099
Fix output number after the restart
