### 2021-02-04 be1a06f
Fix the restart option for the explicit model
Add tests
