### 2021-02-04 654090c
Add restart test
