### 2021-02-05 a33bd63
Fix the wrong output number after restart (#111)
* Fix the wrong output number after restart

* Fix output number after the restart

* Fix the restart option for the explicit model

Add tests

* Add restart test
