+++ examples/fdModel/restart/input_mesh.yaml
@@ -1,0 +1,9 @@
+Output:
+  Path: .
+  Mesh: mesh
+  File_Format: vtu
+Domain: [0.000000e+00, 0.000000e+00, 1.000000e-01, 1.000000e-01]
+Horizon: 2.000000e-03
+Horizon_h_Ratio: 4
+Mesh_Type: uniform_tri
+Compress_Type: zlib
