### 2021-02-05 2626d94
Restart bug (#113)
* Fix the wrong output number after restart

* Fix output number after the restart

* Fix the restart option for the explicit model

Add tests

* Add restart test
