### 2021-02-08 #3020 @whedon
[REVIEW]: NLMech: Implementation of finite difference/meshfreediscretization of nonlocal fracture models
**Submitting author:** @diehlpk (<a href="http://orcid.org/0000-0003-3922-8419">Patrick Diehl</a>)
**Repository:** <a href="https://github.com/nonlocalmodels/NLMech" target ="_blank">https://github.com/nonlocalmodels/NLMech</a>
**Version:**  v0.1.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.16688695.v1" target="_blank">10.6084/m9.figshare.16688695.v1</a>
