+++ .circleci/config.yml
@@ -88,7 +88,7 @@ jobs:
         - run:
             name: Publish code coverage
             command: |
-              if [ "${CIRCLE_BRANCH}" == "master" ]; then
+              if [ "${CIRCLE_BRANCH}" == "main" ]; then
                 cd build
                 coveralls --gcov-options '\-lp' -t ${CTOKEN}
               fi
@@ -108,14 +108,14 @@ jobs:
         - run: 
            name: Deploy to DockerHub
            command: |
-             if [ "${CIRCLE_BRANCH}" == "master" ]; then
+             if [ "${CIRCLE_BRANCH}" == "main" ]; then
                 docker login -u $DOCKER_USER -p $DOCKER_PASS
                 docker push ${IMAGE_NAME}
              fi
         - run: 
            name: Deploy to GitHub
            command: |
-              if [ "${CIRCLE_BRANCH}" == "master" ]; then
+              if [ "${CIRCLE_BRANCH}" == "main" ]; then
                 docker login docker.pkg.github.com --username diehlpk -p ${TOKEN}
                 docker tag ${IMAGE_NAME} docker.pkg.github.com/nonlocalmodels/nlmech/nlmech:latest
                 docker push docker.pkg.github.com/nonlocalmodels/nlmech/nlmech:latest
@@ -138,7 +138,7 @@ jobs:
               make doc
         - run:
             command: |
-                if [ "${CIRCLE_BRANCH}" == "master" ]; then
+                if [ "${CIRCLE_BRANCH}" == "main" ]; then
                   git clone https://github.com/nonlocalmodels/nonlocalmodels.github.io.git
                   cd nonlocalmodels.github.io
                   git fetch
