### 2021-02-20 #3020 @vijaysm
@Kevin-Mattheus-Moerman I can't seem to edit the review checkbox items. Perhaps some permission issues ?
