### 2021-03-23 f104224
The code breaks if the mesh file is not there and does not run with an empty mesh as before
