+++ src/rw/util.cpp
@@ -8,8 +8,6 @@
 
 #include "util.h"
 
-namespace {
-
 std::vector<std::string> get_double_tags() {
   return {"Damage",       "Damage_Phi", "Damage_Z",  "Node_Volume",
           "Nodal_Volume", "Fixity",     "Work_Done", "Strain_Energy"};
@@ -19,7 +17,7 @@ std::vector<std::string> get_point_tags() {
   return {"Displacement", "Velocity", "Force", "Force_Density"};
 }
 
-}  // namespace
+
 
 std::string rw::getDataType(const std::string &data_name) {
   for (const auto &tag : get_double_tags())
@@ -29,4 +27,16 @@ std::string rw::getDataType(const std::string &data_name) {
     if (data_name == tag) return "point";
 
   return "";
+}
+
+bool rw::fileExists(std::string file_name) {
+
+  bool result = boost::filesystem::exists(file_name);
+
+  if (! result) {
+
+    std::cerr << "Error: File " + file_name + " not found!";
+    std::exit(1);
+  } 
+
 }
