+++ src/rw/reader.cpp
@@ -10,6 +10,7 @@
 #include "csv.h"
 #include "mshReader.h"
 #include "vtkReader.h"
+#include "util.h"
 
 void rw::reader::readCsvFile(const std::string &filename, size_t dim,
                              std::vector<util::Point3> *nodes,
@@ -60,6 +61,11 @@ void rw::reader::readVtuFile(const std::string &filename, size_t dim,
                              std::vector<size_t> *enc,
                              std::vector<std::vector<size_t>> *nec,
                              std::vector<double> *volumes, bool is_fd) {
+
+
+  // Check if file exists
+  fileExists(filename);
+
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
   rdr.readMesh(dim, nodes, element_type, num_elem, enc, nec, volumes, is_fd);
@@ -98,6 +104,10 @@ void rw::reader::readVtuFileNodes(const std::string &filename, size_t dim,
 
 bool rw::reader::vtuHasPointData(const std::string &filename,
                                  const std::string &tag) {
+
+  // Check if file exists
+  fileExists(filename);
+
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
   auto has_data = rdr.vtuHasPointData(tag);
@@ -107,6 +117,10 @@ bool rw::reader::vtuHasPointData(const std::string &filename,
 
 bool rw::reader::vtuHasCellData(const std::string &filename,
                                 const std::string &tag) {
+
+  // Check if file exists
+  fileExists(filename);
+
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
   auto has_data = rdr.vtuHasCellData(tag);
@@ -116,6 +130,10 @@ bool rw::reader::vtuHasCellData(const std::string &filename,
 
 std::vector<std::string> rw::reader::readVtuFilePointTags(
     const std::string &filename) {
+
+  // Check if file exists
+  fileExists(filename);
+
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
   auto data = rdr.readVtuFilePointTags();
@@ -125,6 +143,10 @@ std::vector<std::string> rw::reader::readVtuFilePointTags(
 
 std::vector<std::string> rw::reader::readVtuFileCellTags(
     const std::string &filename) {
+  
+  // Check if file exists
+  fileExists(filename);
+  
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
   auto data = rdr.readVtuFileCellTags();
@@ -136,6 +158,9 @@ void rw::reader::readVtuFileCells(const std::string &filename, size_t dim,
                                   size_t &element_type, size_t &num_elem,
                                   std::vector<size_t> *enc,
                                   std::vector<std::vector<size_t>> *nec) {
+  // Check if file exists
+  fileExists(filename);
+  
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
 
@@ -149,6 +174,10 @@ void rw::reader::readVtuFileRestart(const std::string &filename,
                                     std::vector<util::Point3> *u,
                                     std::vector<util::Point3> *v,
                                     const std::vector<util::Point3> *X) {
+  
+  // Check if file exists
+  fileExists(filename);
+
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
   // if displacement is not in input file, use reference coordinate to get
@@ -174,6 +203,10 @@ void rw::reader::readVtuFileRestart(const std::string &filename,
 bool rw::reader::readVtuFilePointData(const std::string &filename,
                                       const std::string &tag,
                                       std::vector<uint8_t> *data) {
+  
+  // Check if file exists
[…]
