+++ src/rw/util.cpp
@@ -17,8 +17,6 @@ std::vector<std::string> get_point_tags() {
   return {"Displacement", "Velocity", "Force", "Force_Density"};
 }
 
-
-
 std::string rw::getDataType(const std::string &data_name) {
   for (const auto &tag : get_double_tags())
     if (data_name == tag) return "double";
@@ -30,13 +28,10 @@ std::string rw::getDataType(const std::string &data_name) {
 }
 
 bool rw::fileExists(std::string file_name) {
-
   bool result = boost::filesystem::exists(file_name);
 
-  if (! result) {
-
+  if (!result) {
     std::cerr << "Error: File " + file_name + " not found!";
     std::exit(1);
-  } 
-
+  }
 }
