+++ src/model/util.cpp
@@ -3,8 +3,6 @@
 
 model::Output::Output(inp::Input *d_input_p, data::DataManager *d_dataManager_p,
                       size_t d_n, double d_time) {
-
-
   std::cout << "Output: time step = " << d_n << "\n";
 
   // write out % completion of simulation at 10% interval
