+++ src/rw/reader.cpp
@@ -61,8 +61,6 @@ void rw::reader::readVtuFile(const std::string &filename, size_t dim,
                              std::vector<size_t> *enc,
                              std::vector<std::vector<size_t>> *nec,
                              std::vector<double> *volumes, bool is_fd) {
-  // Check if file exists
-  fileExists(filename);
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -102,8 +100,6 @@ void rw::reader::readVtuFileNodes(const std::string &filename, size_t dim,
 
 bool rw::reader::vtuHasPointData(const std::string &filename,
                                  const std::string &tag) {
-  // Check if file exists
-  fileExists(filename);
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -114,8 +110,6 @@ bool rw::reader::vtuHasPointData(const std::string &filename,
 
 bool rw::reader::vtuHasCellData(const std::string &filename,
                                 const std::string &tag) {
-  // Check if file exists
-  fileExists(filename);
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -126,8 +120,7 @@ bool rw::reader::vtuHasCellData(const std::string &filename,
 
 std::vector<std::string> rw::reader::readVtuFilePointTags(
     const std::string &filename) {
-  // Check if file exists
-  fileExists(filename);
+
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -138,8 +131,7 @@ std::vector<std::string> rw::reader::readVtuFilePointTags(
 
 std::vector<std::string> rw::reader::readVtuFileCellTags(
     const std::string &filename) {
-  // Check if file exists
-  fileExists(filename);
+
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -152,8 +144,7 @@ void rw::reader::readVtuFileCells(const std::string &filename, size_t dim,
                                   size_t &element_type, size_t &num_elem,
                                   std::vector<size_t> *enc,
                                   std::vector<std::vector<size_t>> *nec) {
-  // Check if file exists
-  fileExists(filename);
+
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -168,8 +159,7 @@ void rw::reader::readVtuFileRestart(const std::string &filename,
                                     std::vector<util::Point3> *u,
                                     std::vector<util::Point3> *v,
                                     const std::vector<util::Point3> *X) {
-  // Check if file exists
-  fileExists(filename);
+  
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -196,8 +186,6 @@ void rw::reader::readVtuFileRestart(const std::string &filename,
 bool rw::reader::readVtuFilePointData(const std::string &filename,
                                       const std::string &tag,
                                       std::vector<uint8_t> *data) {
-  // Check if file exists
-  fileExists(filename);
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -210,8 +198,6 @@ bool rw::reader::readVtuFilePointData(const std::string &filename,
 bool rw::reader::readVtuFilePointData(const std::string &filename,
                                       const std::string &tag,
                                       std::vector<size_t> *data) {
-  // Check if file exists
-  fileExists(filename);
 
   // call vtk reader
   auto rdr = rw::reader::VtkReader(filename);
@@ -224,8 +210,7 @@ bool rw::reader::readVtuFilePointData(const std::string &filename,
 bool rw::reader::readVtuFilePointData(const std::string &filename,
                                       const std::string &tag,
                                       std::vector<int> *data) {
-  // Check if file exists
-  fileExists(filename);
+
 
   // call vtk reader
[…]
