### 2021-03-24 ea09910
Fix missing tests
