### 2021-03-24 957999e
Stop the code if the msh file is not there (#115)
* The code breaks if the mesh file is not there and does not run with an
empty mesh as before

* Code style

* Fix missing tests

* Format
