+++ src/rw/reader.cpp
@@ -10,6 +10,7 @@
 #include "csv.h"
 #include "mshReader.h"
 #include "vtkReader.h"
+#include "util.h"
 
 void rw::reader::readCsvFile(const std::string &filename, size_t dim,
                              std::vector<util::Point3> *nodes,
