### 2021-04-02 42962fd
Add plain stress and plain strain to all models (#121)
* Add plain strain and plain stress to all models

* Format
