+++ src/fe/mesh.h
@@ -354,8 +354,9 @@ private:
    * @param ref_config Base the mesh on reference configuration (obtain
    * reference configuration using point data and displacement data)
    * @param is_centroid_based Specify if we create node at the center of element
+   * @param has_coupling_data Specify if we read the additional data needed for coupling
    */
-  void createData(const std::string &filename, bool ref_config, bool is_centroid_based);
+  void createData(const std::string &filename, bool ref_config, bool is_centroid_based, bool has_coupling_data);
 
   /*!
    * @brief Converts standard fem mesh to particle mesh with nodes at the
@@ -459,6 +460,21 @@ private:
    * volume is computed using the element-node connectivity of the mesh.
    */
   std::vector<double> d_vol;
+
+/*! @brief Vector with the indicies of all nodes with prescribed boundary data
+* 
+* For coupling with other codes, e.g. FEM or PUM, prescribed values for displacement or force
+* are given by the local method and applied to these PD nodes. 
+*/
+  std::vector<size_t> d_prescribed_nodes;
+
+
+/*! @brief Vector with the values of all nodes with prescribed boundary data
+* 
+* For coupling with other codes, e.g. FEM or PUM, prescribed values for displacement or force
+* are given by the local method and applied to these PD nodes. 
+*/
+  std::vector<util::Point3> d_prescribed_values;
 
   /** @}*/
 
