+++ src/fe/mesh.cpp
@@ -75,7 +75,7 @@ fe::Mesh::Mesh(inp::MeshDeck *deck)
     inp::Policy::getInstance()->addToTags(1, "Mesh_d_vol");
 
   // read mesh data from file
-  createData(d_filename, false, deck->d_isCentroidBasedDiscretization);
+  createData(d_filename, false, deck->d_isCentroidBasedDiscretization, deck->d_loadPUMData);
 
   // check if we need to compute mesh size
   if (deck->d_computeMeshSize) computeMeshSize();
@@ -99,7 +99,7 @@ fe::Mesh::Mesh(inp::MeshDeck *deck)
 // Utility functions
 //
 void fe::Mesh::createData(const std::string &filename, bool ref_config,
-                          bool is_centroid_based) {
+                          bool is_centroid_based, bool has_coupling_data) {
   int file_type = -1;
 
   // find the extension of file and call correct reader
@@ -161,6 +161,14 @@ void fe::Mesh::createData(const std::string &filename, bool ref_config,
 
     // check if file has fixity data
     rw::reader::readVtuFilePointData(filename, "Fixity", &d_fix);
+
+    // Read the data for coupling
+    if (has_coupling_data){
+
+       rw::reader::readVtuFilePointData(filename, "Boundary-Layer", &d_prescribed_nodes);
+
+       rw::reader::readVtuFilePointData(filename, "PUM-Displacement", &d_prescribed_values);
+    }
   }
 
   // compute data from mesh data
@@ -408,7 +416,7 @@ void fe::Mesh::readFromFile(inp::MeshDeck *deck, const std::string &filename) {
   d_filename = filename;
 
   // read file
-  createData(filename, false, deck->d_isCentroidBasedDiscretization);
+  createData(filename, false, deck->d_isCentroidBasedDiscretization, deck->d_loadPUMData);
 
   // check if we need to compute mesh size
   if (deck->d_computeMeshSize) computeMeshSize();
