+++ src/fe/mesh.h
@@ -166,10 +166,24 @@ public:
 
   /*!
    * @brief Get the pointer to nodal volume data
-   * @return pointer Pointer to nodal volume data
+   * @return Pointer to nodal volume data
    */
   const std::vector<double> *getNodalVolumesP() const { return &d_vol; };
   std::vector<double> *getNodalVolumesP() { return &d_vol; };
+
+/*!
+*  @brief Get the vector of indicies with prescribed values from the couplign method
+* @return Vector with the indicies
+*/
+  const std::vector<size_t> getPrescribedNodes() const {return d_prescribed_nodes;};
+  std::vector<size_t> getPrescribedNodes() {return d_prescribed_nodes;};
+
+/*!
+*  @brief Get the vector of values of the prescribed nodes from the couplign method
+* @return Vector with the prescribed values
+*/
+  const std::vector<util::Point3> getPrescribedValues() const {return d_prescribed_values;};
+  std::vector<util::Point3> getPrescribedValues() {return d_prescribed_values;};
 
   /*!
    * @brief Return true if node is free
