+++ src/fe/mesh.cpp
@@ -168,6 +168,7 @@ void fe::Mesh::createData(const std::string &filename, bool ref_config,
        rw::reader::readVtuFilePointData(filename, "Boundary-Layer", &d_prescribed_nodes);
 
        rw::reader::readVtuFilePointData(filename, "PUM-Displacement", &d_prescribed_values);
+
     }
   }
 
