### 2021-04-28 19870d8
Update the Readme (#124)
* Switch from master to main in the README

* Fix broken link
