+++ README.md
@@ -1,6 +1,6 @@
 # NLMech [![status](https://joss.theoj.org/papers/271dd66ea91b7fbfdccb4b10a7ba462c/status.svg)](https://joss.theoj.org/papers/271dd66ea91b7fbfdccb4b10a7ba462c)
 
-[![CircleCI](https://circleci.com/gh/nonlocalmodels/NLMech.svg?style=shield)](https://circleci.com/gh/nonlocalmodels/nonlocalheatequation) [![Codacy Badge](https://app.codacy.com/project/badge/Grade/118379d7d745464584b73e9e06f60462)](https://www.codacy.com/gh/nonlocalmodels/NLMech?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=nonlocalmodels/NLMech&amp;utm_campaign=Badge_Grade) [![Coverage Status](https://coveralls.io/repos/github/nonlocalmodels/NLMech/badge.svg?branch=master)](https://coveralls.io/github/nonlocalmodels/NLMech?branch=master)
+[![CircleCI](https://circleci.com/gh/nonlocalmodels/NLMech.svg?style=shield)](https://circleci.com/gh/nonlocalmodels/NLMech) [![Codacy Badge](https://app.codacy.com/project/badge/Grade/118379d7d745464584b73e9e06f60462)](https://www.codacy.com/gh/nonlocalmodels/NLMech?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=nonlocalmodels/NLMech&amp;utm_campaign=Badge_Grade) [![Coverage Status](https://coveralls.io/repos/github/nonlocalmodels/NLMech/badge.svg?branch=main)](https://coveralls.io/github/nonlocalmodels/NLMech?branch=main)
 
 Welcome to NLMech repository. In this project we implement 
 nonlocal fracture theory, referred to as Peridynamics, 
@@ -8,7 +8,7 @@ using both finite element and finite difference discretization.
 
 Logo below has been obtained by running Peridynamic simulation on the logo mesh, [see](https://nonlocalmodels.github.io/examples/fd-logo-soft-material.html).
 
-<p style="text-align:center;"><img src="https://github.com/nonlocalmodels/NLMech/blob/master/assets/logo/logo_sim.png?raw=true" alt="logo" width="400"/></p>
+<p style="text-align:center;"><img src="https://github.com/nonlocalmodels/NLMech/blob/main/assets/logo/logo_sim.png?raw=true" alt="logo" width="400"/></p>
 
 ## Building 
 
@@ -34,11 +34,11 @@ Note that the above provided versions are the minimal rquired versions and we re
 
 We provide following support to build the code
 
-* Bash scripts to build the code on [Ubuntu](https://github.com/nonlocalmodels/buildscripts/tree/master/Ubuntu) and [Fedora](https://github.com/nonlocalmodels/buildscripts/tree/master/Fedora),
-* Docker files to build the code using the [Fedora packages](https://github.com/nonlocalmodels/buildscripts/blob/master/Docker/Fedora) or [build all dependencies](https://github.com/nonlocalmodels/buildscripts/blob/master/Docker/FedoraAll),
+* Bash scripts to build the code on [Ubuntu](https://github.com/nonlocalmodels/buildscripts/tree/main/Ubuntu) and [Fedora](https://github.com/nonlocalmodels/buildscripts/tree/main/Fedora),
+* Docker files to build the code using the [Fedora packages](https://github.com/nonlocalmodels/buildscripts/blob/main/Docker/Fedora) or [build all dependencies](https://github.com/nonlocalmodels/buildscripts/blob/main/Docker/FedoraAll),
[…]
