+++ src/model/quasistatic/QuasiStaticModel.cpp
@@ -171,7 +171,7 @@ void model::QuasiStaticModel<T>::computeForces(bool full) {
 
   // Clear the vector
 
-  hpx::for_loop(hpx::parallel::execution::par, 0, d_nnodes,
+  hpx::for_loop(hpx::execution::par, 0, d_nnodes,
                 [&](boost::uint64_t i) {
                   (*d_dataManager_p->getForceP())[i].d_x = 0.;
                   (*d_dataManager_p->getForceP())[i].d_y = 0.;
@@ -181,7 +181,7 @@ void model::QuasiStaticModel<T>::computeForces(bool full) {
   hpx::lcos::local::mutex m;
 
   hpx::for_loop(
-      hpx::parallel::execution::par, 0, d_nnodes,
+      hpx::execution::par, 0, d_nnodes,
       [&](boost::uint64_t i) {
         util::Point3 force_i = util::Point3();
 
@@ -247,7 +247,7 @@ inline void model::QuasiStaticModel<T>::computePertubatedForces(size_t thread) {
 
   // Clear the vector
 
-  hpx::for_loop(hpx::parallel::execution::par, 0, d_nnodes,
+  hpx::for_loop(hpx::execution::par, 0, d_nnodes,
                 [&](boost::uint64_t i) {
                   (*d_dataManagers[thread]->getForceP())[i].d_x = 0.;
                   (*d_dataManagers[thread]->getForceP())[i].d_y = 0.;
@@ -257,7 +257,7 @@ inline void model::QuasiStaticModel<T>::computePertubatedForces(size_t thread) {
   hpx::lcos::local::mutex m;
 
   hpx::for_loop(
-      hpx::parallel::execution::par, 0, d_nnodes,
+      hpx::execution::par, 0, d_nnodes,
       [&](boost::uint64_t i) {
         util::Point3 force_i = util::Point3();
 
@@ -561,7 +561,7 @@ void model::QuasiStaticModel<T>::solver() {
       auto new_disp = this->newton_step(res);
 
       hpx::for_loop(
-          hpx::parallel::execution::par, 0, d_nnodes, [&](boost::uint64_t i) {
+          hpx::execution::par, 0, d_nnodes, [&](boost::uint64_t i) {
             size_t id = i * dim;
 
             (*d_dataManager_p->getDisplacementP())[i].d_x =
