+++ src/geometry/neighbor.cpp
@@ -67,7 +67,7 @@ geometry::Neighbor::Neighbor(const double &horizon, inp::NeighborDeck *deck,
 #else
 
   auto f = hpx::for_loop(
-      hpx::execution::par(hpx::parallel::execution::task), 0,
+      hpx::execution::par(hpx::execution::task), 0,
       nodes->size(), [this, horizon, nodes](boost::uint64_t i) {
         //  for (size_t i=0; i<nodes->size(); i++) {
         util::Point3 xi = (*nodes)[i];
