+++ tools/pp/src/compute.cpp
@@ -210,7 +210,7 @@ tools::pp::Compute::Compute(const std::string &filename)
 
     if (d_uPlus) {
       auto f2 = hpx::for_loop(
-          hpx::execution::par(hpx::parallel::execution::task), 0,
+          hpx::execution::par(hpx::execution::task), 0,
           d_dataManager_p->getMeshP()->getNumNodes(),
           [this](boost::uint64_t i) {
             d_u[i] += util::Point3(
@@ -896,7 +896,7 @@ void tools::pp::Compute::transformU(rw::writer::Writer *writer) {
                                    util::Point3());
   auto scale = d_currentData->d_transformU_p->d_scale;
   auto f = hpx::for_loop(
-      hpx::execution::par(hpx::parallel::execution::task), 0,
+      hpx::execution::par(hpx::execution::task), 0,
       d_dataManager_p->getMeshP()->getNumNodes(),
       [&u_temp, scale, this](boost::uint64_t i) {
         u_temp[i] = util::Point3(scale * d_u[i].d_x, scale * d_u[i].d_y,
@@ -931,7 +931,7 @@ void tools::pp::Compute::transformV(rw::writer::Writer *writer) {
 
     if (data->d_markVInRectGiven) {
       auto f = hpx::for_loop(
-          hpx::execution::par(hpx::parallel::execution::task), 0,
+          hpx::execution::par(hpx::execution::task), 0,
           d_dataManager_p->getMeshP()->getNumNodes(),
           [&v_mark, data, this](boost::uint64_t i) {
             if (util::geometry::isPointInsideRectangle(
@@ -973,7 +973,7 @@ void tools::pp::Compute::transformV(rw::writer::Writer *writer) {
     if (!data->d_combineMarkV) v_mark = d_v;
 
     auto f = hpx::for_loop(
-        hpx::execution::par(hpx::parallel::execution::task), 0,
+        hpx::execution::par(hpx::execution::task), 0,
         d_dataManager_p->getMeshP()->getNumNodes(),
         [&v_mark, data, this](boost::uint64_t i) {
           auto x = d_dataManager_p->getMeshP()->getNode(i);
@@ -1042,7 +1042,7 @@ void tools::pp::Compute::computeStrain(rw::writer::Writer *writer) {
 
   // compute strain and stress
   auto f = hpx::for_loop(
-      hpx::execution::par(hpx::parallel::execution::task), 0,
+      hpx::execution::par(hpx::execution::task), 0,
       d_dataManager_p->getMeshP()->getNumElements(),
       [&strain, &stress, data, quad, this](boost::uint64_t e) {
         auto ssn = util::SymMatrix3();
@@ -1092,7 +1092,7 @@ void tools::pp::Compute::computeStrain(rw::writer::Writer *writer) {
   if (data->d_magStrainTensor) {
     magS = std::vector<float>(strain.size(), 0.);
     auto f2 = hpx::for_loop(
-        hpx::execution::par(hpx::parallel::execution::task), 0,
+        hpx::execution::par(hpx::execution::task), 0,
         d_dataManager_p->getMeshP()->getNumElements(),
         [&magS, strain, data](boost::uint64_t e) {
           if (data->d_magStrainComp.empty()) {
@@ -1129,7 +1129,7 @@ void tools::pp::Compute::computeStrain(rw::writer::Writer *writer) {
         d_dataManager_p->getMeshP()->getNumElements(), util::Point3());
 
     auto f2 = hpx::for_loop(
-        hpx::execution::par(hpx::parallel::execution::task), 0,
+        hpx::execution::par(hpx::execution::task), 0,
         d_dataManager_p->getMeshP()->getNumElements(),
[…]
