+++ tools/dc/dc.cpp
@@ -5,23 +5,23 @@
 
 #include <algorithm>
 #include <hpx/config.hpp>
-#include <boost/program_options.hpp>
+#include <hpx/modules/program_options.hpp>
 #include <iostream>
 
 #include "dcInclude.h"
 
 int main(int argc, char *argv[]) {
-  boost::program_options::options_description desc("Allowed options");
+  hpx::program_options::options_description desc("Allowed options");
   desc.add_options()("help", "produce help message")(
-      "input-file,i", boost::program_options::value<std::string>(),
+      "input-file,i", hpx::program_options::value<std::string>(),
       "Configuration file")("kind,k",
-                            boost::program_options::value<std::string>(),
+                            hpx::program_options::value<std::string>(),
                             "Kind of simulation");
 
-  boost::program_options::variables_map vm;
-  boost::program_options::store(
-      boost::program_options::parse_command_line(argc, argv, desc), vm);
-  boost::program_options::notify(vm);
+  hpx::program_options::variables_map vm;
+  hpx::program_options::store(
+      hpx::program_options::parse_command_line(argc, argv, desc), vm);
+  hpx::program_options::notify(vm);
 
   if (vm.count("help")) {
     std::cout << desc << "\n";
