+++ .circleci/Docker/Dockerfile
@@ -7,6 +7,8 @@ FROM diehlpk/nonlocalmodels:baseimage
 
 WORKDIR /app
 
+RUN sudo dnf update -y 
+
 RUN git clone  https://github.com/nonlocalmodels/NLMech.git
 
 RUN cd /app/NLMech && mkdir build && cd build && cmake .. -DHPX_IGNORE_COMPILER_COMPATIBILITY=ON -DCMAKE_BUILD_TYPE=Release -DEnable_Tools=ON && make -j 2 && sudo make install
