+++ .circleci/config.yml
@@ -164,8 +164,8 @@ workflows:
       - deploy:
             filters:
               branches:
-                - ignore: gh-pages
-                - only: main
+                 ignore: gh-pages
+                 only: main
             requires:
               - codecoverage
               - documentation
