### 2021-04-29 78c1327
Update the OS before compiling (#127)
* Update the OS before compiling

* Run deploy only on main

* Fix config
