+++ README.md
@@ -14,23 +14,25 @@ Logo below has been obtained by running Peridynamic simulation on the logo mesh,
 
 The code uses 
 
-  * [CMake](https://cmake.org/) 3.10
+  * [CMake](https://cmake.org/) 
 
 as the build system and has following dependencies: 
 
-  * [HPX](https://github.com/STEllAR-GROUP/hpx) 1.5.1
-  * [Blaze](https://bitbucket.org/blaze-lib/blaze/src/master/) 3.8,  
+  * [HPX](https://github.com/STEllAR-GROUP/hpx) 1.6
+  * [Blaze](https://bitbucket.org/blaze-lib/blaze/src/master/),  
   * [Blaze_Iterative](https://github.com/STEllAR-GROUP/BlazeIterative) master, 
-  * [Boost](https://www.boost.org/) 1.73,
-  * [gmsh](https://gmsh.info/) 4.6,
-  * [VTK](https://www.vtk.org) 9.0, and 
-  * [YAML-CPP](https://github.com/jbeder/yaml-cpp) 0.6.3.
+  * [Boost](https://www.boost.org/),
+  * [gmsh](https://gmsh.info/),
+  * [VTK](https://www.vtk.org), and 
+  * [YAML-CPP](https://github.com/jbeder/yaml-cpp).
 
 Following dependencies are optional, but are recommend for large simualtions:
 
-  * [PCL](https://github.com/PointCloudLibrary/pcl) 1.11. 
+  * [PCL](https://github.com/PointCloudLibrary/pcl).
 
-Note that the above provided versions are the minimal rquired versions and we recommend to use these to build the code. 
+Since we use the latest stable [Fedora](https://getfedora.org/) workstation spin for our continiuos testing, we recommend
+these are the minimal required versions to build the latest stable version of the code. For the master branch the same versions
+should work. We strongly recommend to use the same CMake version to build HPX and NLMech.
 
 We provide following support to build the code
 
