+++ README.md
@@ -26,11 +26,11 @@ as the build system and has following dependencies:
   * [VTK](https://www.vtk.org), and 
   * [YAML-CPP](https://github.com/jbeder/yaml-cpp).
 
-Following dependencies are optional, but are recommend for large simualtions:
+Following dependencies are optional, but are recommend for large simulations:
 
   * [PCL](https://github.com/PointCloudLibrary/pcl).
 
-Since we use the latest stable [Fedora](https://getfedora.org/) workstation spin for our continiuos testing, we recommend
+Since we use the latest stable [Fedora](https://getfedora.org/) workstation spin for our continuous testing, we recommend
 these are the minimal required versions to build the latest stable version of the code. For the master branch the same versions
 should work. We strongly recommend to use the same CMake version to build HPX and NLMech.
 
@@ -40,7 +40,7 @@ We provide following support to build the code
 * Docker files to build the code using the [Fedora packages](https://github.com/nonlocalmodels/buildscripts/blob/main/Docker/Fedora) or [build all dependencies](https://github.com/nonlocalmodels/buildscripts/blob/main/Docker/FedoraAll),
 * [Scripts](https://github.com/nonlocalmodels/HPCBuildInfrastructure) to build the code on HPC systems.
 
-If you just interested in running the application, we provide a Doker image with the [latest sucessful build](https://github.com/nonlocalmodels/NLMech/packages/384758) of the main branch. For writing your own 
+If you just interested in running the application, we provide a Doker image with the [latest Successful build](https://github.com/nonlocalmodels/NLMech/packages/384758) of the main branch. For writing your own 
 [YAML](https://docs.ansible.com/ansible/latest/reference_appendices/YAMLSyntax.html) files, we provide a bunch of [examples](https://nonlocalmodels.github.io/examples/).
 
 If you encounter any bug or want to submit a feature request, please generate an [issue](https://github.com/nonlocalmodels/NLMech/issues) on GitHub. 
@@ -52,7 +52,7 @@ The documentation of the the main branch is available [here](https://nonlocalmod
 
 ## Releases
 
-The current stable version is [![GitHub release](https://img.shields.io/github/release/nonlocalmodels/NLMech.svg)](https://GitHub.com/nonlocalmodels/NLMech/releases/) and the current developement happens in the [main branch](https://github.com/nonlocalmodels/NLMech). For more details, we refer to the [Changelog](https://github.com/nonlocalmodels/NLMech/blob/main/CHANGELOG.md) file.
+The current stable version is [![GitHub release](https://img.shields.io/github/release/nonlocalmodels/NLMech.svg)](https://GitHub.com/nonlocalmodels/NLMech/releases/) and the current development happens in the [main branch](https://github.com/nonlocalmodels/NLMech). For more details, we refer to the [Changelog](https://github.com/nonlocalmodels/NLMech/blob/main/CHANGELOG.md) file.
 
 ## Code of conduct
 
@@ -60,7 +60,7 @@ We have adopted a [code of conduct](https://github.com/nonlocalmodels/NLMech/blo
 
 ## Contributing
 
[…]
