+++ README.md
@@ -14,23 +14,25 @@ Logo below has been obtained by running Peridynamic simulation on the logo mesh,
 
 The code uses 
 
-  * [CMake](https://cmake.org/) 3.10
+  * [CMake](https://cmake.org/) 
 
 as the build system and has following dependencies: 
 
-  * [HPX](https://github.com/STEllAR-GROUP/hpx) 1.5.1
-  * [Blaze](https://bitbucket.org/blaze-lib/blaze/src/master/) 3.8,  
+  * [HPX](https://github.com/STEllAR-GROUP/hpx) 1.6
+  * [Blaze](https://bitbucket.org/blaze-lib/blaze/src/master/),  
   * [Blaze_Iterative](https://github.com/STEllAR-GROUP/BlazeIterative) master, 
-  * [Boost](https://www.boost.org/) 1.73,
-  * [gmsh](https://gmsh.info/) 4.6,
-  * [VTK](https://www.vtk.org) 9.0, and 
-  * [YAML-CPP](https://github.com/jbeder/yaml-cpp) 0.6.3.
+  * [Boost](https://www.boost.org/),
+  * [gmsh](https://gmsh.info/),
+  * [VTK](https://www.vtk.org), and 
+  * [YAML-CPP](https://github.com/jbeder/yaml-cpp).
 
-Following dependencies are optional, but are recommend for large simualtions:
+Following dependencies are optional, but are recommend for large simulations:
 
-  * [PCL](https://github.com/PointCloudLibrary/pcl) 1.11. 
+  * [PCL](https://github.com/PointCloudLibrary/pcl).
 
-Note that the above provided versions are the minimal rquired versions and we recommend to use these to build the code. 
+Since we use the latest stable [Fedora](https://getfedora.org/) workstation spin for our continuous testing, we recommend
+these are the minimal required versions to build the latest stable version of the code. For the master branch the same versions
+should work. We strongly recommend to use the same CMake version to build HPX and NLMech.
 
 We provide following support to build the code
 
@@ -38,7 +40,7 @@ We provide following support to build the code
 * Docker files to build the code using the [Fedora packages](https://github.com/nonlocalmodels/buildscripts/blob/main/Docker/Fedora) or [build all dependencies](https://github.com/nonlocalmodels/buildscripts/blob/main/Docker/FedoraAll),
 * [Scripts](https://github.com/nonlocalmodels/HPCBuildInfrastructure) to build the code on HPC systems.
 
-If you just interested in running the application, we provide a Doker image with the [latest sucessful build](https://github.com/nonlocalmodels/NLMech/packages/384758) of the main branch. For writing your own 
+If you just interested in running the application, we provide a Doker image with the [latest Successful build](https://github.com/nonlocalmodels/NLMech/packages/384758) of the main branch. For writing your own 
 [YAML](https://docs.ansible.com/ansible/latest/reference_appendices/YAMLSyntax.html) files, we provide a bunch of [examples](https://nonlocalmodels.github.io/examples/).
 
 If you encounter any bug or want to submit a feature request, please generate an [issue](https://github.com/nonlocalmodels/NLMech/issues) on GitHub. 
@@ -50,7 +52,7 @@ The documentation of the the main branch is available [here](https://nonlocalmod
 
 ## Releases
 
-The current stable version is [![GitHub release](https://img.shields.io/github/release/nonlocalmodels/NLMech.svg)](https://GitHub.com/nonlocalmodels/NLMech/releases/) and the current developement happens in the [main branch](https://github.com/nonlocalmodels/NLMech). For more details, we refer to the [Changelog](https://github.com/nonlocalmodels/NLMech/blob/main/CHANGELOG.md) file.
[…]
