### 2021-05-04 #3020 @vijaysm
@Kevin-Mattheus-Moerman I am in the process of trying to build and run the examples in the repo. Unfortunately, I ran into some snags in getting the TPLs to build properly in my docker environment. I didn't get enough time to dig deeper to resolve the issue yet. But I should be able to get around to all of that this week.
