### 2021-05-05 nonlocalmodels/NLMech#131 @vijaysm
NLMech: Installation of TPL dependencies
This issue page documents some of my review comments for the submission to JOSS [1].

NLMech and its dependency installation/build instructions need to be improved. For example, the following comments apply to some experiments on a stock Ubuntu 18.04 docker image.

1) On Ubuntu 18.04, had to install some dependencies using the following command. It was not clear whether all of this is used directly, or if the HPCInfrastructure build scripts for NLMech configuration will rebuild them to succeed.
```
	$> apt install cmake libboost-dev libyaml-cpp-dev libhwloc-dev libjemalloc-dev
```
It definitely took multiple hops to find a link with the some instructions to install the necessary dependencies. The link in [2] looked like it should have been sufficient to get started. However, using `./build-all.sh Release without-gcc hpx` did not succeed in installing HPX runtime library. The instructions for users to get started could be made simpler using Docker container workflows, and in general would be good if it can work out of the box. And pointers to these instructions should be added to README.md file in NLMech repository, after listing the dependencies and corresponding minimum version numbers as well.

2) To get a bit more specific, CMake fails for HPX on a stock Ubuntu 18.04 docker image. This is due to a minimum of CMake 3.13 requirement by HPX.
```
   	 $> CMake 3.13 or higher is required.  You are running version 3.10.2
```
Upstream minimum dependencies automatically become package dependencies. This is unfortunately true for all build from source workflows. Another potential reason to pre-build a Docker image with dependencies so that users can get started easily with NLMech examples.

3) In general as best practice, it would be good to document the minimum tested/required *versions* to build the TPL dependencies necessary for the library and tools. I see the list of dependencies, but there is no version information provided in README.md.

I will have some more comments as I play with this more and make progress in building NLMech examples locally. Do let me know if you need more details to address some of these concerns.

[1] https://github.com/openjournals/joss-reviews/issues/3020
[2] https://github.com/nonlocalmodels/HPCBuildInfrastructure
