### 2021-05-07 nonlocalmodels/NLMech#131 @diehlpk
> Another potential reason to pre-build a Docker image with dependencies so that users can get started easily with NLMech examples.

We have a link in the README to the Docker image with the latest stable build of the master branch.

> If you just interested in running the application, we provide a Doker image with the latest Successful build of the main branch.

Once we have our first release, after the acceptance of the JOSS paper, we will provide a Docker image with the latest stable release as well.

Please let us know how we can highlight the Docker image more prominently to the user?
