### 2021-05-07 e611a54
Add minimum version of all dependencies (#132)
