### 2021-05-07 nonlocalmodels/NLMech#131 @diehlpk
In addition, we provide

* a Docker file to install the code on Fedora (https://github.com/nonlocalmodels/buildscripts/blob/main/Docker/Fedora)
* a bash script to install the code on Fedora (https://github.com/nonlocalmodels/buildscripts/blob/main/bash/buildFedora.sh)

Note that we use Fedora since we test the code on circle-ci using Fedora and we know that the code should work there.
