### 2021-05-07 nonlocalmodels/NLMech#131 @vijaysm
Thanks for the prompt response. Appreciate the quick updates to README to link the scripts and docker images.

> Please let us know how we can highlight the Docker image more prominently to the user?

It would help to add a "Getting Started" section in README that explains steps to run a simple example in NLMech. This can use the Docker-based workflow with either Fedora or Ubuntu. With the containerized solutions, the user does not have to worry about TPL installations as the environment is hardened with testing. This can accelerate user adoption of the library.

> Thanks, @vijaysm for your comments. In addition to what Patrick said, we will further improve the documentation and simplify the building of the library by providing a single script file that installs all dependencies (other than those available via apt-get) and NLMech code.

Excellent. I'll play with the Fedora container and add more comments soon.
