### 2021-05-07 fb707db
Add installation tutorial
