### 2021-05-07 fec3a05
Use new authentication method for coveralls.io
