### 2021-05-07 nonlocalmodels/NLMech#131 @diehlpk
> @diehlpk I will add the ubuntu 18.04 and 20.04 docker base images and corresponding files to build the code. We can check code build on circle-ci using both ubuntu and fedora.

Cool, please add them to the buildscript repo.
