### 2021-05-07 nonlocalmodels/NLMech#131 @vijaysm
Just a quick note that I can't seem to pull the docker image as per instructions.
```
$ > docker pull docker.pkg.github.com/nonlocalmodels/nlmech/nlmech:latest
Error response from daemon: Head https://docker.pkg.github.com/v2/nonlocalmodels/nlmech/nlmech/manifests/latest: no basic auth credentials
```
Please update the [documentation page](https://github.com/nonlocalmodels/NLMech/packages/384758) to point to the right docker image.
