### 2021-05-07 nonlocalmodels/NLMech#131 @diehlpk
> Just a quick note that I can't seem to pull the docker image as per instructions.
>
> ```
> $ > docker pull docker.pkg.github.com/nonlocalmodels/nlmech/nlmech:latest
> Error response from daemon: Head https://docker.pkg.github.com/v2/nonlocalmodels/nlmech/nlmech/manifests/latest: no basic auth credentials
> ```
>
> Please update the [documentation page](https://github.com/nonlocalmodels/NLMech/packages/384758) to point to the right docker image.

Ok, that is a bug with the Github Docker and you need togin with your Github account to even pull a public Docker image, see here for more details: https://github.community/t/docker-pull-from-public-github-package-registry-fail-with-no-basic-auth-credentials-error/16358/14

I will point to the Docker Hub image, since that image should be public available. Since I am always connected to Github, I could not check this.
