### 2021-05-07 nonlocalmodels/NLMech#131 @diehlpk
@vijaysm Would you mind to test this Docker image

https://hub.docker.com/r/diehlpk/nlmech/tags?page=1&ordering=last_updated
