### 2021-05-07 2e9a89e
Switch to the Docker Hub image
