+++ docs/paper.bib
@@ -103,7 +103,8 @@
 @mastersthesis{diehl2012implementierung,
   title={Implementierung eines Peridynamik-Verfahrens auf GPU},
   author={Diehl, Patrick},
-  year={2012}
+  year={2012},
+  doi={10.18419/opus-2879}
 }
 
 @incollection{diehl2015efficient,
