+++ docs/paper.bib
@@ -10,256 +10,225 @@
 }
 
 @article{geuzaine2009gmsh,
-	title={Gmsh: A 3-D finite element mesh generator with built-in pre-and post-processing facilities},
-	author={Geuzaine, Christophe and Remacle, Jean-Fran{\c{c}}ois},
-	journal={International journal for numerical methods in engineering},
-	volume={79},
-	number={11},
-	pages={1309--1331},
-	year={2009},
-	publisher={Wiley Online Library}
+	author = {Geuzaine, Christophe and Remacle, Jean-Fran{\c c}ois},
+	journal = {International journal for numerical methods in engineering},
+	number = {11},
+	pages = {1309--1331},
+	publisher = {Wiley Online Library},
+	title = {Gmsh: A 3-D finite element mesh generator with built-in pre-and post-processing facilities},
+	volume = {79},
+	year = {2009}
 }
 
 @article{jha2019numerical,
-  title={Numerical convergence of finite difference approximations for state based peridynamic fracture models},
-  author={Jha, Prashant K and Lipton, Robert},
-  journal={Computer Methods in Applied Mechanics and Engineering},
-  volume={351},
-  pages={184--225},
-  year={2019},
-  publisher={Elsevier}
+	author = {Jha, Prashant K and Lipton, Robert},
+	journal = {Computer Methods in Applied Mechanics and Engineering},
+	pages = {184--225},
+	publisher = {Elsevier},
+	title = {Numerical convergence of finite difference approximations for state based peridynamic fracture models},
+	volume = {351},
+	year = {2019}
 }
 
 @article{lipton2019complex,
-  title={Complex fracture nucleation and evolution with nonlocal elastodynamics},
-  author={Lipton, Robert P and Lehoucq, Richard B and Jha, Prashant K},
-  journal={Journal of Peridynamics and Nonlocal Modeling},
-  volume={1},
-  number={2},
-  pages={122--130},
-  year={2019},
-  publisher={Springer}
+	author = {Lipton, Robert P and Lehoucq, Richard B and Jha, Prashant K},
+	journal = {Journal of Peridynamics and Nonlocal Modeling},
+	number = {2},
+	pages = {122--130},
+	publisher = {Springer},
+	title = {Complex fracture nucleation and evolution with nonlocal elastodynamics},
+	volume = {1},
+	year = {2019}
 }
 
 @article{lipton2018free,
-  title={Free damage propagation with memory},
-  author={Lipton, Robert and Said, Eyad and Jha, Prashant},
-  journal={Journal of Elasticity},
-  volume={133},
-  number={2},
-  pages={129--153},
-  year={2018},
-  publisher={Springer}
+	author = {Lipton, Robert and Said, Eyad and Jha, Prashant},
+	journal = {Journal of Elasticity},
+	number = {2},
+	pages = {129--153},
+	publisher = {Springer},
+	title = {Free damage propagation with memory},
+	volume = {133},
+	year = {2018}
 }
 
 @article{jha2020kinetic,
-  title={Kinetic relations and local energy balance for LEFM from a nonlocal peridynamic model},
-  author={Jha, Prashant K and Lipton, Robert P},
-  journal={International Journal of Fracture},
-  pages={1--15},
-  year={2020},
-  publisher={Springer}
+	author = {Jha, Prashant K and Lipton, Robert P},
+	journal = {International Journal of Fracture},
+	pages = {1--15},
+	publisher = {Springer},
+	title = {Kinetic relations and local energy balance for LEFM from a nonlocal peridynamic model},
+	year = {2020}
 }
 
 @article{littlewood2015roadmap,
-  title={Roadmap for peridynamic software implementation},
-  author={Littlewood, David J},
[…]
