+++ docs/paper.bib
@@ -215,13 +215,19 @@
 	year = {2019}
 }
 
+
 @article{diehl2018implementation,
-	author = {Diehl, Patrick and Jha, Prashant K and Kaiser, Hartmut and Lipton, Robert and Levesque, Martin},
-	journal = {arXiv},
-	pages = {arXiv--1806},
-	title = {Implementation of Peridynamics utilizing HPX--the C++ standard library for parallelism and concurrency},
-	year = {2018}
+  title={An asynchronous and task-based implementation of peridynamics utilizing HPX—the C++ standard library for parallelism and concurrency},
+  author={Diehl, Patrick and Jha, Prashant K and Kaiser, Hartmut and Lipton, Robert and L{\'e}vesque, Martin},
+  journal={SN Applied Sciences},
+  volume={2},
+  number={12},
+  pages={1--21},
+  year={2020},
+  publisher={Springer}
 }
+
+
 
 @article{Diehl2020,
 	author = {Diehl, Patrick},
