+++ docs/paper.bib
@@ -6,7 +6,8 @@
 	publisher = {SIAM},
 	title = {Numerical analysis of nonlocal fracture models in Holder space},
 	volume = {56},
-	year = {2018}
+	year = {2018},
+    doi={10.1137/17m1112236}
 }
 
 @article{geuzaine2009gmsh,
@@ -17,7 +18,8 @@
 	publisher = {Wiley Online Library},
 	title = {Gmsh: A 3-D finite element mesh generator with built-in pre-and post-processing facilities},
 	volume = {79},
-	year = {2009}
+	year = {2009},
+    doi={10.1002/nme.2579}
 }
 
 @article{jha2019numerical,
@@ -27,7 +29,8 @@
 	publisher = {Elsevier},
 	title = {Numerical convergence of finite difference approximations for state based peridynamic fracture models},
 	volume = {351},
-	year = {2019}
+	year = {2019},
+    doi={10.1016/j.cma.2019.03.024}
 }
 
 @article{lipton2019complex,
@@ -38,7 +41,8 @@
 	publisher = {Springer},
 	title = {Complex fracture nucleation and evolution with nonlocal elastodynamics},
 	volume = {1},
-	year = {2019}
+	year = {2019},
+    doi={10.1007/s42102-019-00010-0}
 }
 
 @article{lipton2018free,
@@ -49,7 +53,8 @@
 	publisher = {Springer},
 	title = {Free damage propagation with memory},
 	volume = {133},
-	year = {2018}
+	year = {2018},
+    doi={10.1007/s10659-018-9672-7}
 }
 
 @article{jha2020kinetic,
@@ -58,14 +63,16 @@
 	pages = {1--15},
 	publisher = {Springer},
 	title = {Kinetic relations and local energy balance for LEFM from a nonlocal peridynamic model},
-	year = {2020}
+	year = {2020},
+    doi={10.1007/s10704-020-00480-0}
 }
 
 @article{littlewood2015roadmap,
 	author = {Littlewood, David J},
 	journal = {SAND Report, Sandia National Laboratories, Albuquerque, NM and Livermore, CA},
 	title = {Roadmap for peridynamic software implementation},
-	year = {2015}
+	year = {2015},
+    doi={10.2172/1226115}
 }
 
 @article{parks2008implementing,
@@ -76,7 +83,8 @@
 	publisher = {Elsevier},
 	title = {Implementing peridynamics within a molecular dynamics code},
 	volume = {179},
-	year = {2008}
+	year = {2008},
+    doi={10.1016/j.cpc.2008.06.011}
 }
 
 @article{ren20173d,
@@ -86,7 +94,8 @@
 	publisher = {Elsevier},
 	title = {A 3D discontinuous Galerkin finite element method with the bond-based peridynamics model for dynamic brittle failure analysis},
 	volume = {99},
-	year = {2017}
+	year = {2017},
+    doi={10.1016/j.ijimpeng.2016.09.003}
 }
 
 @article{mossaiby2017opencl,
@@ -97,7 +106,8 @@
 	publisher = {Elsevier},
 	title = {OpenCL implementation of a high performance 3D Peridynamic model on graphics accelerators},
 	volume = {74},
-	year = {2017}
+	year = {2017},
+    doi={10.1016/j.camwa.2017.06.045}
 }
 
 @mastersthesis{diehl2012implementierung,
@@ -113,18 +123,20 @@
 	pages = {81--95},
 	publisher = {Springer},
 	title = {Efficient neighbor search for particle methods on GPUs},
-	year = {2015}
+	year = {2015},
+    doi={10.1007/978-3-319-06898-5_5}
 }
 
 @article{silling2007peridynamic,
[…]
