+++ docs/paper.bib
@@ -136,7 +136,7 @@
 	title = {Peridynamic states and constitutive modeling},
 	volume = {88},
 	year = {2007},
-    doi={10.1007/s10659-007-9125-1]
+    doi={10.1007/s10659-007-9125-1}
 }
 
 @article{lipton2014dynamic,
