### 2021-05-11 #3020 @diehlpk
@chennachaos The references are fixed and all DOIs were added.
