### 2021-05-17 #3020 @chennachaos
Hi @diehlpk. Thanks for fixing the issues with references.

After a careful look at the documentation, I have the following comments.

1.) Although the important details seem to be there, the documentation is spanned over two different websites. Installation instructions are in the new documentation and the examples and other details are in the old site. Some details are on both sites. I recommend combining these two.

2.) One of the links under "Please cite following publications, if you use the code" is broken.

3.) The installation instructions are quite clear. However, the functionality documentation is severely lacking. You need to add a page or two describing the fundamental equations that the library solves, and a brief overview of the method(s) used.

4.) The documentation contains some good numerical examples. But the description of these numerical examples needs improvement. For each example, add the problem background, the objective, problem setup and the parameters to be studied etc. Also, describe different figures presented in the results section of examples provided.

5.) Add one or two additional numerical examples with some plots comparing the results with reference values so that the reader gets a sense of accuracy right away.

6.) The examples provided at the moment are too simple. Please add some examples with complex geometries. Add at least one 3D example.

7.) Add documentation on preparing the input files in the .yml format, together with the explanations of different blocks and the parameters used in each block in the input file.

8.) Explain/discuss the details needed to run the parallel versions of examples.

9.) Based on the results provided in the [paper](https://arxiv.org/pdf/1806.06917.pdf), it looks like the code has so far been tested only on a small number of CPUs, a maximum of 8. Moreover, the parallel framework used is multi-threading.
Could you please elaborate on the parallel framework adapted for this library, HPX, especially its suitability and scalability for distributed computing for simulating large-scale 3D problems? How easy, or difficult, it is to switch from multithreading parallelism to full distributed parallelism across multiple nodes using HPX?
