### 2021-05-18 #3020 @diehlpk
> 9.) Based on the results provided in the [paper](https://arxiv.org/pdf/1806.06917.pdf), it looks like the code has so far been tested only on a small number of CPUs, a maximum of 8. Moreover, the parallel framework used is multi-threading.
> Could you please elaborate on the parallel framework adapted for this library, HPX, especially its suitability and scalability for distributed computing for simulating large-scale 3D problems? How easy, or difficult, it is to switch from multithreading parallelism to full distributed parallelism across multiple nodes using HPX?

Sure, we had some Google Summer of Code intern last summer, and he implemented a distributed version of a non-local models, the non-local diffusion, in a separate repo

https://github.com/nonlocalmodels/nonlocalheatequation

However, this code hast not the advanced functionality as NLMech. This work will be published this year in a IDPDS workshop proceeding and the preprint is available here

https://arxiv.org/abs/2102.03819

So, having a distributed version of NLMech is possible based on this previous work. As usual the developers have no funding to extend this code to a distributed version. Currently, we only have funding to work on the crack and fracture simulations, develop new models, and run simulations to validate some numerical analysis.

Having a distributed implementation in HPX is on our todo list and will hopefully have some funding soon.
