### 2021-05-18 #3020 @diehlpk
> 6.) The examples provided at the moment are too simple. Please add some examples with complex geometries. Add at least one 3D example.

With respect to the complex geometry, the code's logo

https://raw.githubusercontent.com/nonlocalmodels/NLMech/main/assets/logo/logo_sim.png

geometry was used in these examples

- https://nonlocalmodels.github.io/examples/fd-logo-soft-material.html
- https://nonlocalmodels.github.io/examples/fd-logo-soft-material-2.html

and colorized with the obtained damage. The user can use gmsh to design complex geometries and just the msh file as an input.

If this geometry is not complex enough, what kind of geometry you have in mind?
