### 2021-05-18 #3020 @diehlpk
> 5.) Add one or two additional numerical examples with some plots comparing the results with reference values so that the reader gets a sense of accuracy right away.

We have several papers which we will mention in the documentation and make more clear where to find the accuracy results.
