### 2021-05-18 #3020 @diehlpk
> 2.) One of the links under "Please cite following publications, if you use the code" is broken.

@chennachaos Could you please be more specific, I could not find the broken link.
