### 2021-05-18 3ff503f
Add some example on how running the code
