### 2021-05-18 #3020 @diehlpk
> 8.) Explain/discuss the details needed to run the parallel versions of examples.

I added some explanation of the command line options to run the code in parallel here

https://github.com/nonlocalmodels/NLMech#running-nlmech
