### 2021-05-26 d0cc851
Add example how to run the docker image and run one 1D examples
