### 2021-05-26 #3020 @diehlpk
@vijaysm I added some example to the README.md here

https://github.com/nonlocalmodels/NLMech/pull/140

how to

- pull the docker image
- run the docker image using the bash
- and run one of the test case in 1D to take not too long to finish in the Docker image.
