### 2021-05-26 #3020 @diehlpk
`3.) The installation instructions are quite clear. However, the functionality documentation is severely lacking. You need to add a page or two describing the fundamental equations that the library solves, and a brief overview of the method(s) used.`

@chennachaos We have all this in the methods paper

Diehl, P., Jha, P. K., Kaiser, H., Lipton, R., & Lévesque, M. (2020). An asynchronous and task-based implementation of peridynamics utilizing HPX—the C++ standard library for parallelism and concurrency. SN Applied Sciences, 2(12), 1-21.

Jha, PK, Lipton R. “Numerical convergence of finite difference approximations for state based peridynamic fracture models.” Computer Methods in Applied Mechanics and Engineering, 1 July 2019, 351(1), 184 - 225. Link

I am happy to mention this in the documentation, however, I do not think that it is necessary to add all these equations to the documentation.
