### 2021-05-27 #3020 @diehlpk
> 1.) Although the important details seem to be there, the documentation is spanned over two different websites. Installation instructions are in the new documentation and the examples and other details are in the old site. Some details are on both sites. I recommend combining these two.

@chennachaos This is on purpose to only have the light-weight documentation, e.g. installation, running the code, and details about the configuration files, in the code repo (https://github.com/nonlocalmodels/NLMech). The heavy documentation, e.g. images for the examples, are outsourced to the webpage's (https://nonlocalmodels.github.io/) repo (https://github.com/nonlocalmodels/nonlocalmodels.github.io). The doxygen documentation is published there as well (https://nonlocalmodels.github.io/documentation/).

We do not think that images should be inside the code repo, especially if you will submit the code as a Fedora or Ubuntu package to their package manager it is not nice if you have images shipped with the source code.
