### 2021-05-27 f07f023
Add documentation for the YAML files for the mesh generation
