### 2021-05-27 #3020 @diehlpk
> 7.) Add documentation on preparing the input files in the .yml format, together with the explanations of different blocks and the parameters used in each block in the input file.

@chennachaos We started to add the documentation for the configuration files. I added the mesh generation files here

https://github.com/nonlocalmodels/NLMech/pull/141

We will add more for the simulation, however, that will take some time.
