### 2021-06-07 #3020 @vijaysm
@diehlpk Some review comments on the paper.

- Line 19: Replace "I.e." with "i.e.,"
- Line 21: Missing commas: "radius $\delta$, centered at x, where"
- Line 44: Correct this: "Figure Figure 2"
- Figure 2: Caption. Fix "Adapted"
- It may help to talk about the theory and the software implementation of it separately, rather than mixing them under the same section. Specifically, look at lines 53-56. My suggestion here is to combine this paragraph with the "Applications" section and make it "Software Implementation and Applications" for better clarity.
- Line 74: Missing comma: "from an open source perspective, only two"
- Line 79: No need for a bracket here: "we refer to Diehl et al., 2020."
