### 2021-06-07 74b1cd0
Minor grammar fixes
