### 2021-06-09 40aed77
Update the qs 2d example input file for the new format
