### 2021-06-09 nonlocalmodels/NLMech#131 @diehlpk
@vijaysm Just for your information, in #142 one tested the new Docker image, and it worked for him. Except the failed test.
