### 2021-06-13 #3020 @chennachaos
Thank you @diehlpk for your responses!

2) The link to "10.1007/s42452-020-03784-x" in the paper  "Please cite following publications, if you use the code" is broken.
