### 2021-06-13 #3020 @chennachaos
1.) This is fine in that case.

2.) The link to "10.1007/s42452-020-03784-x" in the paper "Please cite following publications, if you use the code" is broken.

3.) I understand that the equations are provided in the paper but the online documentation should contain some basic equations, if not all, so that the reader can easily relate to what is being solved without needing to refer to the papers all the time.

4.) Not addressed.

5.)  I don't think it is sufficient to refer to the papers for the accuracy studies. The documentation must describe a few examples; this should include some comparisons with literature (experiment/numerical/analytical). You may use the same graphs or other comparisons from your papers.

6.) I think these 2D geometries are complex enough. It is understandable that 3D problems are computationally more demanding and that it will require considerable effort to parallelism the code for distributed computing. However, I believe that the code should be able to solve some simple 3D problems, even without parallelisation.

7.) This (documentation on preparing input files) needs to be completed before the submission can be recommended for publication. Otherwise, the users will struggle to use the library.

8.) This is good.

9.) Thanks for the clarification! It is understandable that parallelization for distributed memory architectures takes time. Until then, please mention in the documentation that the current capability is limited to thread-based parallelism.
