### 2021-06-14 #3020 @diehlpk
> 9.) Thanks for the clarification! It is understandable that parallelization for distributed memory architectures takes time. Until then, please mention in the documentation that the current capability is limited to thread-based parallelism.

@chennachaos That was added to the Readme.md and the main page of the documentation.
