+++ docs/conf.doxy.in
@@ -813,6 +813,7 @@ INPUT                  = ../src/main.cpp \
                          ../docs/content/cmake-options.md \
                          ../docs/content/install-instructions.md \
                          ../docs/content/yaml-options.md \
+                         ../docs/content/equations.md \
                          ../README.md 
                          
 
