### 2021-06-14 37c342a
Add new try with the eqauations
