+++ docs/content/equations.md
@@ -1,14 +1,3 @@
-
-<script type="text/x-mathjax-config">
-    MathJax.Hub.Config({
-      tex2jax: {
-        skipTags: ['script', 'noscript', 'style', 'textarea', 'pre'],
-        inlineMath: [['$','$']]
-      }
-    });
-  </script>
-  <script src="https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" type="text/javascript"></script>
-
 # Brief overview of the equations
 
 The governing equations and the discretization is briefly introduced on this page. For more details we
@@ -16,31 +5,31 @@ refer to the following two references [1,2].
 
 ## Governing equations
 
-Unlike classical continuum mechanics, where the internal force is written in terms of the stress, in PD, the internal force at a given material point is due to the sum of the pairwise forces of the neighboring points. i.e. the force in PD is expressed as the integral of the pairwise force density between the given point and another point in the neighborhood. Neighborhood of point `$x$` is typically taken as the ball of radius `$\delta$`, centered at $x$, where `$\delta$` is the nonlocal length scale and is referred to as horizon. PD is often divided in two classes: bond-based and state-based models. In bond-based models, the two material points interact via a pairwise force law and the forces between the material points do not depend on the deformation state of surrounding points. In contrast, in the state-based models the volumetric deformation in the neighborhood of two points plays a role in the pariwise force. The governing equation of motion for the bond-based PD [3] reads as
[…]
