+++ docs/content/equations.md
@@ -18,8 +18,7 @@ Here <a href="https://www.codecogs.com/eqnedit.php?latex=\dpi{120}&space;\varrho
 
 ## Discretization 
 
-Currently, the library supports finite difference (or more generally meshfree) discretization. Using the triangulation of arbitrary domain, the library can create a meshfree discretization. The library is equipped with necessary modules, such as FE elements and quadrature integration rules, for finite element discretization of PD. Next, we briefly discuss the finite difference/meshfree discretization of PD. Figure 1 shows the domain <a href="https://www.codecogs.com/eqnedit.php?latex=\dpi{120}&space;D" target="_blank"><img src="https://latex.codecogs.com/gif.latex?\dpi{120}&space;D" title="D" /></a> discretized with the nodes <a href="https://www.codecogs.com/eqnedit.php?latex=\dpi{120}&space;X&space;=&space;\{&space;X_i&space;\in&space;\mathbb{R}^3&space;\vert&space;i=1,\ldots,n\}" target="_blank"><img src="https://latex.codecogs.com/gif.latex?\dpi{120}&space;X&space;=&space;\{&space;X_i&space;\in&space;\mathbb{R}^3&space;\vert&space;i=1,\ldots,n\}" title="X = \{ X_i \in \mathbb{R}^3 \vert i=1,\ldots,n\}" /></a>. Each node <a href="https://www.codecogs.com/eqnedit.php?latex=\dpi{120}&space;X_i" target="_blank"><img src="https://latex.codecogs.com/gif.latex?\dpi{120}&space;X_i" title="X_i" /></a> represents a small area/volume denoted by <a href="https://www.codecogs.com/eqnedit.php?latex=\dpi{120}&space;V_i" target="_blank"><img src="https://latex.codecogs.com/gif.latex?\dpi{120}&space;V_i" title="V_i" /></a>. In PD, as previously mentioned, each point <a href="https://www.codecogs.com/eqnedit.php?latex=\dpi{120}&space;X_i" target="_blank"><img src="https://latex.codecogs.com/gif.latex?\dpi{120}&space;X_i" title="X_i" /></a> interacts with neighboring points in ball (discrete) <a href="https://www.codecogs.com/eqnedit.php?latex=\dpi{120}&space;B_\delta(X_i)&space;=&space;\{X_j:&space;\vert&space;X_i&space;-&space;X_j|&space;<&space;\delta&space;\}" target="_blank"><img src="https://latex.codecogs.com/gif.latex?\dpi{120}&space;B_\delta(X_i)&space;=&space;\{X_j:&space;\vert&space;X_i&space;-&space;X_j|&space;<&space;\delta&space;\}" title="B_\delta(X_i) = \{X_j: \vert X_i - X_j| < \delta \}" /></a>. 
-
[…]
