### 2021-06-14 #3020 @diehlpk
`3.) I understand that the equations are provided in the paper but the online documentation should contain some basic equations, if not all, so that the reader can easily relate to what is being solved without needing to refer to the papers all the time.`

@chennachaos I added the basic equations showed in the JOSS paper to the documentation and referenced them in the Readme.md and the index.html of the documentation.

See https://github.com/nonlocalmodels/NLMech/pull/144
