### 2021-06-14 #3020 @diehlpk
> 5.) I don't think it is sufficient to refer to the papers for the accuracy studies. The documentation must describe a few examples; this should include some comparisons with literature (experiment/numerical/analytical). You may use the same graphs or other comparisons from your papers.

We can not easily reuse the images from the papers, since the copyright was transferred to the publisher. So we would need to be careful here. In addition, the documentation is generated using doxygen and the support of equations is limited there.

We have two examples where we showed that we recover properties from linear elasticity:

* https://nonlocalmodels.github.io/examples/qs-1d-properties.html
* https://nonlocalmodels.github.io/examples/qs-2d-elastic.html

All other benchmarks are very technical, and I do not think there is any benefit to copy the results from the paper to the documentation, since all papers are available as preprints and the reader can read there the exact same information.
