### 2021-06-15 5847adb
Add all common yaml options
