### 2021-06-15 #3020 @diehlpk
> 7.) This (documentation on preparing input files) needs to be completed before the submission can be recommended for publication. Otherwise, the users will struggle to use the library.

@chennachaos All the common attributes which are essential to get started with the code are added. Meaning, all tested and user-friendly attributes are documented. However, experimental and under current development are not yet there, but will later be added once the features are stable.

https://github.com/nonlocalmodels/NLMech/pull/145
