### 2021-06-15 #3020 @diehlpk
> 6.) I think these 2D geometries are complex enough. It is understandable that 3D problems are computationally more demanding and that it will require considerable effort to parallelism the code for distributed computing. However, I believe that the code should be able to solve some simple 3D problems, even without parallelisation.

@chennachaos

Here, all data structures and algorithms are implemented to do 3D simulations. However, the geometrical functions, like applying boundary conditions not on a line or rectangle, are missing for 3D. We intend to add these features once we start the parallel implementation and add the load balancing as described in

https://arxiv.org/abs/2102.03819

From my experience, 3D simulations need fully distributed computations to run some complex geometry since the amount of nodes gets really large.
