### 2021-06-16 ccac215
Add all common yaml attributes to the documentation (JOSS review) (#145)
* Add all common yaml options

* Spellcheck
