+++ docs/content/yaml-options.md
@@ -45,6 +45,195 @@ The tag `Domain` describes the simulation domain using the following attributes:
 
 For more complex geometries, the [Gmsh](https://gmsh.info/) tool is recommended. Further examples are available [here](https://nonlocalmodels.github.io/examples/).
 
+## Simulation
+
+Here, all tags and attributes of the YAML files are listed and described. A detailed set of examples is available [here](https://nonlocalmodels.github.io/examples/).
+
+### Model deck
+
+Example of a model deck:
+
+```yaml
+Model:
+  Dimension: 2
+  Discretization_Type:
+    Spatial: finite_difference
+    Time: velocity_verlet
+  Final_Time: 0.001000
+  Time_Steps: 50000
+  Horizon: 0.002000
+  Horizon_h_Ratio: 4
+```
+
+The tag `Model` describes the model with following attributes:
+
+* `Dimension` Defines the dimension of the problem, e.g. one-dimensional 
+* `Discretization_Type` Describes the discretization in time and space
+  * `Time` Defines if a central difference scheme `central_difference` or a velocity verlet `velocity_verlet` scheme is used for the discretization in time
+  * `Spatial` Defines if a finite difference `finite_difference` scheme is used
+* `Horizon` Defines the horizon
+* `Horizon_h_Ratio` Defines the ratio of nodes with the distance of h in the horizon
+* `Mesh_Size` Defines the mesh size h
+* `Final_Time` Defines the final time 
+* `Time_Steps` Defines the amount of time steps
+
+### Restart
+
+Example of a `Restart` deck:
+
+```yaml
+Restart:
+  File: out/output_3.vtu
+  Step: 1500
+```
+
+The tag `Restart` describes a restart of the simulation using following attributes:
+
+* `File` The path and file name of the last Successful simulation step
+* `Step` The time step to restart from
+
+One example of the restart of one simulation is available [here](https://nonlocalmodels.github.io/examples/restart.html).
+
+### Mesh
+
+Example of a `Mesh` deck:
+
+```yaml
+Mesh:
+  File: mesh.vtu
+  Keep_Element_Conn: true
+```
+
+The tag `Mesh` describes the mesh of the simulation using following attributes:
+
+* `File` Path and file name to the mesh either in the gmsh or vtu file format.
+* `Keep_Element_Conn` Keep the mesh information available
+
+### Output
+
+Example of a `Output` deck
+
+```yaml
+Output:
+  Path: out_top_bottom_const/
+  Tags:
+    - Displacement
+    - Velocity
+    - Force
+    - Damage_Z
+  Output_Interval: 500
+  Compress_Type: zlib
+  Perform_FE_Out: true
+```
+
+The tag `Output` describes the mesh of the simulation using following attributes:
+
+* `File_Format` Specifies if the gmsh or vtu file format is used
+* `Path` Defines the path were the output is written to
+* `Compress_Type` Defines the compression type for the vtu file (`zlib` or `ascii`)
+* `Output_Interval` Defines the interval a output file is written
+* `Tag` Defines the field appended to the out file. The mesh is always added and all other simulation data needs to be listed.
+* `Perform_FE_Out` Store the mesh information in the output
+
+### Boundary conditions
+
+#### Displacement boundary conditions
+
+Example of a `Displacement_BC` deck
+
+```yaml
+Displacement_BC:
+  Sets: 1
+  Set_1:
+    Location:
+      Rectangle: [0.000000e+00, 9.900000e-02, 1.000000e-01, 1.000000e-01]
+    Direction: [1,2]
+    Time_Function:
+      Type: constant
+      Parameters:
+        - 0.0
+    Spatial_Function:
+      Type: constant
+      Parameter: [1]
+```
+
+
[…]
