### 2021-06-16 #3020 @diehlpk
> 4.) The documentation contains some good numerical examples. But the description of these numerical examples needs improvement. For each example, add the problem background, the objective, problem setup and the parameters to be studied etc. Also, describe different figures presented in the results section of examples provided.

@chennachaos

- We made clear that the two examples with the code do not have any background. They were just there to make the code's logo.
- Most of these examples were used in publications, and we just presented this examples here. We added the corresponding publications and linked the YAML files, so users can experience the full file and adapt it for theirusage.
