### 2021-07-01 nonlocalmodels/NLMech#131 @vijaysm
For the 2-D example, I do still get a segfault error like the one reported in Issue #142 . Output stack trace below.
```
{stack-trace}: 4 frames:
0x7fb484682a20  : /lib64/libpthread.so.0(+0x13a20) [0x7fb484682a20] in /lib64/libpthread.so.0
0x7fb48354f0a3  : hpx::threads::coroutines::detail::coroutine_impl::operator()() [0xc3] in /lib64/libhpx_core.so
0x7fb48354e72f  : /lib64/libhpx_core.so(+0xa072f) [0x7fb48354e72f] in /lib64/libhpx_core.so
{what}: Segmentation fault
```
I just pulled the [NLMech docker image](https://hub.docker.com/r/diehlpk/nlmech). Should I use the image from a different location ? I see the recent update was 15 days back.
