### 2021-07-02 nonlocalmodels/NLMech#131 @diehlpk
> Couple of comments on the README instructions on the Docker image usage. Excerpt from README.
>
> ```
> cd examples/qsModel/1D
> # Generate the mesh file
> /app/NLMech/build/bin/mesh -i input_mesh.yaml -d 1
> # Run the simulation
> /app/NLMech/build/bin/NLMech -i input.yaml --hpx:threads=2
> ```
>
>     * Here, perhaps make the instruction explicit to `cd /app/NLMech/examples/qsModel/1D`.
>
>     * The command `/app/NLMech/build/bin/NLMech -i input.yaml --hpx:threads=2` runs fine. But using `--hpx:threads=1` is much faster on my image, and `--hpx:threads=4` is slower than both. This is counter-intuitive. I ran each command multiple times to verify the timing discrepancy. Sample runtimes below.
>
>       1. **--hpx:threads=1**: Time elapsed = 0.219245 sec
>       2. **--hpx:threads=2**: Time elapsed = 0.577724 sec
>       3. **--hpx:threads=4**: Time elapsed = 1.07554 sec
>
>     * Please rebuild HPX in the docker image to see if that fixes the above problem. However, I want to confirm that I can run the 1D example without issues.
>
>
> I'll have more comments soon.

This es exactly what to expect, since the 1d problem is too small to gain any benefit by using more threads at all. What you see is the overhead by adding more threads to the computation. This problem is only for validation and not really for scaling runs. You would need to increase the amount of points there to see scaling. Therefore, we never published scaling results in 1D.
