### 2021-07-02 190c59c
Update the path for the docker example
