+++ examples/fdModel/logo/README.md
@@ -1,4 +1,4 @@
-### Sinusoidal wrt space and linear in time loading on 2-d Logo
+### Example files for the code's logo
 
 Fore more details, see [here](https://nonlocalmodels.github.io/examples/fd-logo-soft-material.html)
 
