### 2021-07-02 nonlocalmodels/NLMech#131 @diehlpk
>     * Here, perhaps make the instruction explicit to `cd /app/NLMech/examples/qsModel/1D`.

I started to work on that and will do some pull request soon.
