### 2021-07-02 nonlocalmodels/NLMech#131 @diehlpk
> For the 2-D example, I do still get a segfault error like the one reported in Issue #142 . Output stack trace below.
>
> ```
> {stack-trace}: 4 frames:
> 0x7fb484682a20  : /lib64/libpthread.so.0(+0x13a20) [0x7fb484682a20] in /lib64/libpthread.so.0
> 0x7fb48354f0a3  : hpx::threads::coroutines::detail::coroutine_impl::operator()() [0xc3] in /lib64/libhpx_core.so
> 0x7fb48354e72f  : /lib64/libhpx_core.so(+0xa072f) [0x7fb48354e72f] in /lib64/libhpx_core.so
> {what}: Segmentation fault
> ```
>
> I just pulled the [NLMech docker image](https://hub.docker.com/r/diehlpk/nlmech). Should I use the image from a different location ? I see the recent update was 15 days back.

@vijaysm Have you run `docker pull image` before you tried that?

```
[root@c9861c720d71 2D]# /app/NLMech/build/bin/NLMech -i input.yaml --hpx:threads=2
/app/NLMech/build/bin/NLMech (Version 0.1.0)
QSModel: Initializing high level objects.
QSModel: Creating mesh.
Mesh: Reading mesh.
Number of nodes = 289
QSModel: Initializing displacement loading object.
QSModel: Initializing force loading object.
QSModel: Creating neighbor list.
QSModel: Initializing material object.
QSModel: Initializing Jacobian matrix (578x578).
Output: time step = 0
```

I tested the latest docker image right now and it seems to work for me?
