### 2021-07-02 nonlocalmodels/NLMech#131 @vijaysm
AFAIK, I did have the latest. Nevertheless, I deleted my local containers for your repo, and did the following:

- `docker pull diehlpk/nlmech` gives me the hash 4022c564068999b6e0961658ba207ef184ab48b3679803fb522184af8d657162.

- I then generated the mesh with `/app/NLMech/build/bin/mesh -i input_mesh.yaml -d 2`
- Next I run the application with `/app/NLMech/build/bin/NLMech -i input.yaml --hpx:threads=2`. This now works cleanly and does not segfault like I originally saw before. So perhaps my image was still outdated and didn't pull the latest version as I thought it did. Example output below..
 ```[root@cc5e6a4763dc 2D]# /app/NLMech/build/bin/NLMech -i input.yaml --hpx:threads=2
/app/NLMech/build/bin/NLMech (Version 0.1.0)
QSModel: Initializing high level objects.
QSModel: Creating mesh.
Mesh: Reading mesh.
Number of nodes = 289
QSModel: Initializing displacement loading object.
QSModel: Initializing force loading object.
QSModel: Creating neighbor list.
QSModel: Initializing material object.
QSModel: Initializing Jacobian matrix (578x578).
Output: time step = 0
It: 0 Res: 1030.78
It: 1 Res: 316.847
It: 2 Res: 151.912
```
