### 2021-07-02 nonlocalmodels/NLMech#131 @diehlpk
> > This is exactly what to expect, since the 1d problem is too small (33 discrete PD nodes) to gain any benefit by using more threads at all. What you see is the overhead by adding more threads to the computation. This problem is only for validation and not really for scaling runs. You would need to increase the amount of points there to see scaling. Therefore, we never published scaling results in 1D.
>
> Understood. Nevertheless, this is non-intuitive as your users will expect a different behavior.
>
> My suggestion would then be to perhaps use the 2-D example as the default in README when using `hpx:threads` arguments so that the performance improvement is more apparent with more threads.

The 2D quasi-static example takes rather long, and we wanted to avoid that the users has to wait for such a long time. I would suggest making a comment that for scaling the 2D example should be used.
