### 2021-07-02 nonlocalmodels/NLMech#131 @vijaysm
> The 2D quasi-static example takes rather long, and we wanted to avoid that the users has to wait for such a long time. I would suggest making a comment that for scaling the 2D example should be used.

Faire enough. Please update the README accordingly.
