### 2021-07-11 #3020 @diehlpk
> Hi @diehlpk,
> I am having difficulties installing the library. The problem is with the **Blaze** and **Blaze_Iterative** libraries.
> I was hoping that I could make use of the scripts provided but there are no files at the link *Build Dependencies and Code on Ubuntu*.
> I work with Ubuntu OS.

@chennachaos are you using the HPC Buildinfrastructure Scripts as shown in the documentation?

These scripts download and install all dependencies. Note that both libraries are headed only libraries.

Can you post any error message?
