### 2021-07-12 #3020 @diehlpk
@chennachaos Have you seen this tutorial to compile the code

https://nonlocalmodels.github.io/documentation/md_content_install_instructions.html
