### 2021-07-12 #3020 @vijaysm
@Kevin-Mattheus-Moerman I'm in a similar stage where my installation efforts for the TPL dependencies fail using the provided scripts. But I've been messaging @diehlpk who has provided a docker container with these pre-installed. I have tested the 1D and 2D examples using that container, and I am satisfied with the setup.

However, I'd like to follow the instructions to build NLMech again from scratch if possible. I have not yet tried building on a Fedora base system (I used Ubuntu as well) since that's what the scripts above use.
