### 2021-07-12 #3020 @diehlpk
@vijaysm and @chennachaos could you please post some error messages and I might, can help you.

My institution's cluster runs on CentOS and my desktop on Fedora, so the scripts are not really tested. What Ubuntu version are you  using? I can try to install the code using Docker.
