### 2021-07-12 #3020 @diehlpk
@vijaysm and @chennachaos I pulled the Ubuntu latest Docker image from docker.io this morning.

And I had to install the following packages using apt-get

* apt-get update
* apt-get install build-essential
* apt-get install git
* apt-get install wget
* apt-get install cmake
* apt-get install openssl
* apt-get install libssl-dev
* apt-get install libblas-dev liblapack-dev
* apt-get install autoconf
* apt-get install freeglut3-dev

After that I run

git clone https://github.com/nonlocalmodels/HPCBuildInfrastructure.git
cd HPCBuildInfrastructure

and executed the following commands to build all dependencies:

* ./build-all.sh Release without-gcc cmake
* ./build-all.sh Release without-gcc hwloc
* ./build-all.sh Release without-gcc jemalloc
* ./build-all.sh Release without-gcc blaze
* ./build-all.sh Release without-gcc blazeIterative
* ./build-all.sh Release without-gcc boost
* ./build-all.sh Release without-gcc yamlcpp
* ./build-all.sh Release without-gcc hpx
*  ./build-all.sh Release without-gcc vtk

All of them compiled for me out of the box, however, I fixed to small issue related to Docker in the build scripts.

I have to finish something else today, and will look into NLMech later this week.

It would really help, if you could post any error messages you have seen, since I could not see any error with the latest Ubunut verison.
