### 2021-07-12 #3020 @vijaysm
Thanks for the details @diehlpk and reverifying the instructions. I had my
installation failures in some notes. I'll dig it up. I'm happy with the
docker solution however, and if the script you had is more suitable for
Fedora that should be fine as well. But I see you tested on Ubuntu now, and
so I'll follow the new instructions.

One point that I explicitly remember was that I needed to install openssl
and ssl-dev, which was not mentioned in the documentation. A dependency of
HPX perhaps, but it's the curse of stacks.

Also, I was trying to use system installed cmake as I did not want to
install it from source. This had some issues as well. Why do you need to
build cmake from source? If you need the most up to date version, please
specify that in documentation or README also. Otherwise, I'd recommend
simplify instructions with just specifying - depends on CMake (>=3.1.0) for
example.

I should have some more info for you later today, but it probably is
outdated since the tests were done in May.

On Mon., Jul. 12, 2021, 16:25 Patrick Diehl, ***@***.***>
wrote:

> @vijaysm <https://github.com/vijaysm> and @chennachaos
> <https://github.com/chennachaos> I pulled the Ubuntu latest Docker image
> from docker.io this morning.
>
> And I had to install the following packages using apt-get
>
>    - apt-get update
>    - apt-get install build-essential
>    - apt-get install git
>    - apt-get install wget
>    - apt-get install cmake
>    - apt-get install openssl
>    - apt-get install libssl-dev
>    - apt-get install libblas-dev liblapack-dev
>    - apt-get install autoconf
>    - apt-get install freeglut3-dev
>
> After that I run
>
> git clone https://github.com/nonlocalmodels/HPCBuildInfrastructure.git
> cd HPCBuildInfrastructure
>
> and executed the following commands to build all dependencies:
>
>    - ./build-all.sh Release without-gcc cmake
>    - ./build-all.sh Release without-gcc hwloc
>    - ./build-all.sh Release without-gcc jemalloc
>    - ./build-all.sh Release without-gcc blaze
>    - ./build-all.sh Release without-gcc blazeIterative
>    - ./build-all.sh Release without-gcc boost
>    - ./build-all.sh Release without-gcc yamlcpp
>    - ./build-all.sh Release without-gcc hpx
>    - ./build-all.sh Release without-gcc vtk
>
> All of them compiled for me out of the box, however, I fixed to small
> issue related to Docker in the build scripts.
>
> I have to finish something else today, and will look into NLMech later
> this week.
>
> It would really help, if you could post any error messages you have seen,
> since I could not see any error with the latest Ubunut verison.
>
