### 2021-07-12 #3020 @diehlpk
>  Also, I was trying to use system installed cmake as I did not want to install it from source. This had some issues as well. Why do you need to build cmake from source? If you need the most up to date version, please specify that in documentation or README also. Otherwise, I'd recommend simplify instructions with just specifying - depends on CMake (>=3.1.0) for example.

HPX relies on a recent CMake version, and all codes on top of HPX need to use the same CMake version. I will update the minimal version on the README.md
