### 2021-07-12 #3020 @diehlpk
> One point that I explicitly remember was that I needed to install openssl and ssl-dev, which was not mentioned in the documentation. A dependency of HPX perhaps, but it's the curse of stacks.

Yes, that is a huge issue with things like openssl, autoconf, and so on. That is really operating system specific and there is no silver bullet to list all of them.

From my experience, these issues are for all C++ applications. I compiled HPX on many supercomputers, e.g. Summit, Piz Daint, and Cori. And we would not find some generic list of packages needed, since all clusters had a different software stack.
