### 2021-07-12 #3020 @diehlpk
@chennachaos and @vijaysm

I added some new script to compile the code on Ubuntu

https://github.com/nonlocalmodels/buildscripts/blob/main/bash/buildAll_Ubuntu.sh

This might help.
