### 2021-07-13 051c27c
Docker (#146)
* Update the path for the docker example

* Update
