### 2021-07-15 #3020 @chennachaos
Hi @diehlpk, the other dependencies you installed using **apt-get install** needs to be added to the list of dependencies on the installation page.

A major issue with the current support for building the library is that it seems to be built explicitly for HPC machines and, therefore, it is not compatible with the installation on personal computers. Use of *module* commands in the build scripts prevent installations on personal computers which usually do not have *module* management software installed on them.

Also, the idea of installing all the libraries separately is not good. This requires a significant amount of resources for compiling and building huge libraries like **CMAKE** and **VTK**.

There should be a provision to make use of libraries that are installed.
