### 2021-07-15 #3020 @diehlpk
> Use of _module_ commands in the build scripts prevent installations on personal computers which usually do not have _module_ management software installed on them.

@chennachaos That is not correct, since the scripts generates the modules for you, but there is not need to use them to run the code.
