### 2021-07-15 #3020 @diehlpk
> Also, the idea of installing all the libraries separately is not good. This requires a significant amount of resources for compiling and building huge libraries like **CMAKE** and **VTK**.

@chennachaos You can use your own cmake, however, HPX relies on a very recent one. Especially in Ubuntu the CMake version is not recent enough.
