### 2021-07-15 #3020 @diehlpk
> Also, the idea of installing all the libraries separately is not good. This requires a significant amount of resources for compiling and building huge libraries like **CMAKE** and **VTK**.

@chennachaos Note that you are using the tool chain for HPX system to build the code. If you want to use it on your desktop computer you might want to follow this instructions

https://github.com/nonlocalmodels/NLMech#compiling-library

There is no need to compile all the dependencies, like I have shown here for Fedora

https://github.com/nonlocalmodels/buildscripts/blob/main/bash/buildFedora.sh

CMake checks for installed libraries on your system or you can provide the path if the library is not installed in the system directories.

Since we have some hpx package on Fedora, there is no need to compile HPX. On other systems you have to build HPX, but you can use boost, hwloc, and most of the dependencies from the package manager.

I for example use for development the instruction on Fedora in the script above.
