### 2021-07-15 #3020 @diehlpk
> Hi @diehlpk, the other dependencies you installed using **apt-get install** needs to be added to the list of dependencies on the installation page.

@chennachaos That is a tricky question, because this are not dependencies of NLMech. These are dependencies of the dependencies of NLMech. So I am not sure if it is really necessary to add a full list of dependencies of the dependencies. Especially were should we stop to track down the dependencies?

Note that these dependencies are specific for the server version of Ubuntu. If I use the desktop version, there is no need to install openssl.
