### 2021-07-16 #3020 @vijaysm
> @chennachaos That is a tricky question, because this are not dependencies of NLMech. These are dependencies of the dependencies of NLMech. So I am not sure if it is really necessary to add a full list of dependencies of the dependencies. Especially were should we stop to track down the dependencies?

Unfortunately, this is the curse of dependency stacks. It is important to make clear the overall minimum set of dependency list (and yes, you inherit dependencies of your dependencies as well) that you need to build NLMech consistently. If you want to restrict usage to Fedora and Ubuntu, then fine tune the scripts for the platforms and document that installations in other platforms are untested/unverified.

I raised this issue initially, but I am a bit more convinced that there is a viable usage path with the Docker images that I have already used to verify the NLMech examples.

> Note that these dependencies are specific for the server version of Ubuntu. If I use the desktop version, there is no need to install openssl.

Just to clarify, I had to install openssl on the Ubuntu desktop (?) version since I was just using the docker image for 20.04.

> Since there is not a package manager, like pip or anaconda in python, I do believe it is really difficult to provide on script to install the code on various operating systems. We could provide specific scripts for each OS, but that would be too much work to keep them updated.

One suggestion here is to try to use a package manager like spack to simplify dependency builds instead of using custom written bash scripts which are prone to failure depending on the base system and available pre-installed dependencies. With spack, there are recipes you could create in order to build NLMech and TPLs consistently. That being said, I have not checked if they support HPX and other dependencies that you require already.
