### 2021-07-16 #3020 @diehlpk
> One suggestion here is to try to use a package manager like spack to simplify dependency builds instead of using custom written bash scripts which are prone to failure depending on the base system and available pre-installed dependencies. With spack, there are recipes you could create in order to build NLMech and TPLs consistently. That being said, I have not checked if they support HPX and other dependencies that you require already.

I think we have HPX available in spack and maybe in easyinstall. However,  blaze and blaze_iterative are not available in spack and easytinsall. I do not have the capacity to add blaze and easyinstall to them right now. Maybe in the future, we will work on that but that will not happen in the near future.
