### 2021-07-16 #3020 @diehlpk
> > @chennachaos That is a tricky question, because this are not dependencies of NLMech. These are dependencies of the dependencies of NLMech. So I am not sure if it is really necessary to add a full list of dependencies of the dependencies. Especially were should we stop to track down the dependencies?
>
> Unfortunately, this is the curse of dependency stacks. It is important to make clear the overall minimum set of dependency list (and yes, you inherit dependencies of your dependencies as well) that you need to build NLMech consistently. If you want to restrict usage to Fedora and Ubuntu, then fine tune the scripts for the platforms and document that installations in other platforms are untested/unverified.

@vijaysm and @chennachaos I added them to the Readme.md and also listed the command to install them using apt or dnf.

In addition, I made some comment that some dependencies can be installed using the system packages. I added the example to install them using dnf or apt as well.
