### 2021-07-16 #3020 @vijaysm
@diehlpk thanks for updating the documentation.

With regards to spack, looks like they do have blaze [1] if that was a bottleneck before.

[1] https://spack.readthedocs.io/en/latest/package_list.html#blaze
