### 2021-07-16 #3020 @diehlpk
> @diehlpk thanks for updating the documentation.
>
> With regards to spack, looks like they do have blaze [1] if that was a bottleneck before.
>
> [1] https://spack.readthedocs.io/en/latest/package_list.html#blaze

Ok, but we still miss Blaze_Iterative. I agree, having the code in Spack would be really, however, I do not have the capacity to add Blaze_Iterative to Spack and maintain the package.
