### 2021-07-16 #3020 @vijaysm
> Ok, but we still miss Blaze_Iterative.

Sure fair enough. But something to consider for future - and perhaps a motivation to convince the Blaze_Iterative devs to maintain the recipe as well for broader usage. Not really relevant to this review and so I don't want to diverge too much from topic.

Overall, I have preferred the Dockerized solution to getting started with NLMech, since this avoids the need for a user to build and install all these complex TPLs. You can add the build all TPLs from source as part of the developer guide or for advanced users since it would take a non-trivial amount of time to get it built, on some user machine.

My review for NLMech is almost complete. I will take a look to confirm that there are no other pending comments I have written down elsewhere. Thanks for all the quick responses @diehlpk
