### 2021-07-16 #3020 @diehlpk
> Overall, I have preferred the Dockerized solution to getting started with NLMech, since this avoids the need for a user to build and install all these complex TPLs. You can add the build all TPLs from source as part of the developer guide or for advanced users since it would take a non-trivial amount of time to get it built, on some user machine.

Yes, I totally agree that we have some complex TPLs, and it is not easy to install the code. As you said, we provide the docker image for people who just want to play with the code.
