### 2021-07-20 nonlocalmodels/NLMech#131 @vijaysm
I'm satisfied with all the changes. Thanks for the prompt responses @diehlpk
