### 2021-07-20 #3020 @vijaysm
@Kevin-Mattheus-Moerman My review of NLMech is complete. Apart from some hiccups in the installation process from source, I do not have any other pending issues as @diehlpk has promptly addressed all the concerns I raised here, and at [NLMech issue #131](https://github.com/nonlocalmodels/NLMech/issues/131). I recommend acceptance of the submission once all the other reviews are in. Sorry about the longer than average delay in getting my review done.
