### 2021-07-26 #3020 @chennachaos
> > Use of _module_ commands in the build scripts prevent installations on personal computers which usually do not have _module_ management software installed on them.
>
> @chennachaos That is not correct, since the scripts generates the modules for you, but there is not need to use them to run the code.

Hi @diehlpk, I don't understand how the installation can be carried out if *module* command is not available.

I would like to finish the review by the end of this week. Do you have any other method that I can follow to install and test the code on my PC? I have Ubuntu 20.04.
