### 2021-07-26 #3020 @diehlpk
> Hi @diehlpk, I don't understand how the installation can be carried out if _module_ command is not available.
>
> I would like to finish the review by the end of this week. Do you have any other method that I can follow to install and test the code on my PC? I have Ubuntu 20.04.

Hey @chennachaos you can just run the scripts, as I showed above. They do not use the module command at all. The scripts just write modules file, you can use on clusters, but they are optional.

I think if you follow these instructions on your Ubuntu, you can compile the code:

And I had to install the following packages using apt-get

* apt-get update
* apt-get install build-essential
* apt-get install git
* apt-get install wget
* apt-get install cmake
* apt-get install libssl-dev
* apt-get install libblas-dev liblapack-dev
* apt-get install autoconf
* apt-get install freeglut3-dev

After that I run

git clone https://github.com/nonlocalmodels/HPCBuildInfrastructure.git
cd HPCBuildInfrastructure

and executed the following commands to build all dependencies:

* ./build-all.sh Release without-gcc cmake
* ./build-all.sh Release without-gcc hwloc
* ./build-all.sh Release without-gcc jemalloc
* ./build-all.sh Release without-gcc blaze
* ./build-all.sh Release without-gcc blazeIterative
* ./build-all.sh Release without-gcc boost
* ./build-all.sh Release without-gcc yamlcpp
* ./build-all.sh Release without-gcc hpx
*  ./build-all.sh Release without-gcc vtk

I think this is what you need to do to compile the code. If you install vtk or boost using ap-get, you can skip the build-all.sh for them and CMake will use the one installed on the system.
