### 2021-07-29 #3020 @chennachaos
Hi @diehlpk. I managed to install all the dependencies using the commands you provided. Thanks!

I encountered some errors while building the NLMech library. The error is due to CMake being unable to find the path to the HPX library/package.
How can I fix this?
