### 2021-07-29 #3020 @diehlpk
> Hi @diehlpk. I managed to install all the dependencies using the commands you provided. Thanks!
>
> I encountered some errors while building the NLMech library. The error is due to CMake being unable to find the path to the HPX library/package.
> How can I fix this?
@chennachaos I added a new script

./build-all.sh Release without-gcc nl

and you should please use this script to compile the code. This script will set the path to all installed libraries using the previous scripts.
