### 2021-07-30 #3020 @diehlpk
> I pulled the new file. I am getting the same error even with **./build-all.sh Release without-gcc nl**.
> The problem is in the installation of dependencies. In particular, the directories and other environment variables do not seem to be set properly.

Can you please post the error to help me to check what is going wrong?
