### 2021-08-03 #3020 @chennachaos
@diehlpk, see below for the out of ls commands.

chenna@chenna-XPS-L501X:~/Documents/HPCBuildInfrastructure$ ls -lah build
total 48K
drwxrwxr-x 12 chenna chenna 4.0K Jul 27 13:43 .
drwxrwxr-x  6 chenna chenna 4.0K Jul 29 17:26 ..
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:03 blaze
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:04 blaze-iterative
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:08 boost
drwxrwxr-x  6 chenna chenna 4.0K Jul 27 10:54 cmake
drwxrwxr-x  7 chenna chenna 4.0K Jul 27 11:51 hpx
drwxrwxr-x  9 chenna chenna 4.0K Jul 27 11:00 hwloc
drwxr-xr-x  6 chenna chenna 4.0K Jul 27 11:02 jemalloc
drwxrwxr-x 10 chenna chenna 4.0K Jul 27 13:43 modules
drwxrwxr-x  6 chenna chenna 4.0K Jul 27 13:43 vtk
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:14 yamlcpp
chenna@chenna-XPS-L501X:~/Documents/HPCBuildInfrastructure$ ls -lah build/hpx/lib/cmake/HPX/
total 664K
drwxr-xr-x 5 chenna chenna 4.0K Jul 27 11:51 .
drwxrwxr-x 3 chenna chenna 4.0K Jul 27 11:51 ..
-rw-r--r-- 1 chenna chenna 2.1K Jul 27 11:20 FindAmplifier.cmake
-rw-r--r-- 1 chenna chenna  564 Jul 27 11:20 FindBreathe.cmake
-rw-r--r-- 1 chenna chenna 2.4K Jul 27 11:20 FindBZip2.cmake
-rw-r--r-- 1 chenna chenna 3.0K Jul 27 11:20 FindGooglePerftools.cmake
-rw-r--r-- 1 chenna chenna 2.2K Jul 27 11:20 FindHwloc.cmake
-rw-r--r-- 1 chenna chenna 1.8K Jul 27 11:20 FindIbverbs.cmake
-rw-r--r-- 1 chenna chenna 1.1K Jul 27 11:20 FindIB_VERBS.cmake
-rw-r--r-- 1 chenna chenna 2.7K Jul 27 11:20 FindJemalloc.cmake
-rw-r--r-- 1 chenna chenna 1.4K Jul 27 11:20 FindLibfabric.cmake
-rw-r--r-- 1 chenna chenna 1.9K Jul 27 11:20 FindLibSigSegv.cmake
-rw-r--r-- 1 chenna chenna 1.6K Jul 27 11:20 FindMSR.cmake
-rw-r--r-- 1 chenna chenna 1.7K Jul 27 11:20 FindOrangeFS.cmake
-rw-r--r-- 1 chenna chenna 2.2K Jul 27 11:20 FindPAPI.cmake
-rw-r--r-- 1 chenna chenna 1.7K Jul 27 11:20 FindPMI.cmake
-rw-r--r-- 1 chenna chenna 1.7K Jul 27 11:20 FindQThreads.cmake
-rw-r--r-- 1 chenna chenna 1.4K Jul 27 11:20 FindRdmacm.cmake
-rw-r--r-- 1 chenna chenna 1.1K Jul 27 11:20 FindRDMA_CM.cmake
-rw-r--r-- 1 chenna chenna 1.4K Jul 27 11:20 FindSnappy.cmake
-rw-r--r-- 1 chenna chenna  525 Jul 27 11:20 FindSphinx.cmake
-rw-r--r-- 1 chenna chenna 1.5K Jul 27 11:20 FindTBB.cmake
-rw-r--r-- 1 chenna chenna 2.4K Jul 27 11:20 FindTBBmalloc.cmake
-rw-r--r-- 1 chenna chenna 1.9K Jul 27 11:20 FindTCMalloc.cmake
-rw-r--r-- 1 chenna chenna 1.1K Jul 27 11:20 FindValgrind.cmake
-rw-r--r-- 1 chenna chenna 9.5K Jul 27 11:20 GitExternal.cmake
-rw-r--r-- 1 chenna chenna 7.0K Jul 27 11:20 HPX_AddCompileFlag.cmake
-rw-r--r-- 1 chenna chenna 5.5K Jul 27 11:20 HPX_AddCompileTest.cmake
-rw-r--r-- 1 chenna chenna 9.7K Jul 27 11:20 HPX_AddComponent.cmake
-rw-r--r-- 1 chenna chenna  19K Jul 27 11:20 HPX_AddConfigTest.cmake
-rw-r--r-- 1 chenna chenna 6.3K Jul 27 11:20 HPX_AddDefinitions.cmake
-rw-r--r-- 1 chenna chenna 7.0K Jul 27 11:20 HPX_AddExecutable.cmake
-rw-r--r-- 1 chenna chenna 8.8K Jul 27 11:20 HPX_AddLibrary.cmake
-rw-r--r-- 1 chenna chenna 2.6K Jul 27 11:20 HPX_AddLibraryHeaders.cmake
-rw-r--r-- 1 chenna chenna 2.4K Jul 27 11:20 HPX_AddLibrarySources.cmake
-rw-r--r-- 1 chenna chenna 3.0K Jul 27 11:20 HPX_AddLinkFlag.cmake
-rw-r--r-- 1 chenna chenna  11K Jul 27 11:20 HPX_AddModule.cmake
-rw-r--r-- 1 chenna chenna 2.1K Jul 27 11:20 HPX_AddParcelport.cmake
-rw-r--r-- 1 chenna chenna 1.2K Jul 27 11:20 HPX_AddPseudoDependencies.cmake
-rw-r--r-- 1 chenna chenna  754 Jul 27 11:20 HPX_AddPseudoTarget.cmake
-rw-r--r-- 1 chenna chenna 1.7K Jul 27 11:20 HPX_AddSourceGroup.cmake
-rw-r--r-- 1 chenna chenna 7.6K Jul 27 11:20 HPX_AddTest.cmake
-rw-r--r-- 1 chenna chenna  521 Jul 27 11:20 HPX_AppendProperty.cmake
-rw-r--r-- 1 chenna chenna  24K Jul 27 11:21 HPXCacheVariables.cmake
-rw-r--r-- 1 chenna chenna 2.3K Jul 27 11:20 HPX_CheckCXXStandard.cmake
-rw-r--r-- 1 chenna chenna 1.5K Jul 27 11:20 HPX_CompilerFlagsTargets.cmake
-rw-r--r-- 1 chenna chenna 3.9K Jul 27 11:21 HPXConfig.cmake
-rw-r--r-- 1 chenna chenna 1.4K Jul 27 11:21 HPXConfigVersion.cmake
-rw-r--r-- 1 chenna chenna 1.3K Jul 27 11:20 HPX_CreateSymbolicLink.cmake
-rw-r--r-- 1 chenna chenna  322 Jul 27 11:20 HPX_CXXOverrides.cmake
-rw-r--r-- 1 chenna chenna 1.7K Jul 27 11:20 HPX_Documentation.cmake
-rw-r--r-- 1 chenna chenna  874 Jul 27 11:20 HPX_ExportTargets.cmake
-rw-r--r-- 1 chenna chenna  587 Jul 27 11:20 HPX_ForceOutOfTreeBuild.cmake
-rw-r--r-- 1 chenna chenna  417 Jul 27 11:20 HPX_FortranCompiler.cmake
-rw-r--r-- 1 chenna chenna  368 Jul 27 11:20 HPX_FortranOverrides.cmake
-rw-r--r-- 1 chenna chenna 1.3K Jul 27 11:20 HPX_ForwardCacheVariables.cmake
-rw-r--r-- 1 chenna chenna 5.3K Jul 27 11:20 HPX_GeneratePackage.cmake
-rw-r--r-- 1 chenna chenna  13K Jul 27 11:20 HPX_GeneratePackageUtils.cmake
-rw-r--r-- 1 chenna chenna 1.3K Jul 27 11:20 HPX_GitCommit.cmake
-rw-r--r-- 1 chenna chenna  765 Jul 27 11:20 HPX_HandleComponentDependencies.cmake
-rw-r--r-- 1 chenna chenna  328 Jul 27 11:20 HPX_Include.cmake
-rw-r--r-- 1 chenna chenna  61K Jul 27 11:21 HPXInternalTargets.cmake
-rw-r--r-- 1 chenna chenna  43K Jul 27 11:21 HPXInternalTargets-release.cmake
-rw-r--r-- 1 chenna chenna 5.2K Jul 27 11:21 HPXMacros.cmake
-rw-r--r-- 1 chenna chenna 2.7K Jul 27 11:20 HPX_Message.cmake
-rw-r--r-- 1 chenna chenna 3.4K Jul 27 11:20 HPX_Option.cmake
-rw-r--r-- 1 chenna chenna 3.6K Jul 27 11:20 HPX_PerformCxxFeatureTests.cmake
-rw-r--r-- 1 chenna chenna 3.2K Jul 27 11:20 HPX_PrintSummary.cmake
-rw-r--r-- 1 chenna chenna  357 Jul 27 11:20 HPX_SetCMakePolicy.cmake
-rw-r--r-- 1 chenna chenna  463 Jul 27 11:20 HPX_SetFullRPATH.cmake
-rw-r--r-- 1 chenna chenna  980 Jul 27 11:20 HPX_SetLibName.cmake
-rw-r--r-- 1 chenna chenna 1.6K Jul 27 11:20 HPX_SetOutputPaths.cmake
-rw-r--r-- 1 chenna chenna 1.2K Jul 27 11:20 HPX_SetPlatform.cmake
-rw-r--r-- 1 chenna chenna 5.4K Jul 27 11:20 HPX_SetupAllocator.cmake
-rw-r--r-- 1 chenna chenna 2.7K Jul 27 11:20 HPX_SetupApex.cmake
-rw-r--r-- 1 chenna chenna 4.5K Jul 27 11:20 HPX_SetupBoost.cmake
-rw-r--r-- 1 chenna chenna 1.1K Jul 27 11:20 HPX_SetupBoostFilesystem.cmake
-rw-r--r-- 1 chenna chenna  642 Jul 27 11:20 HPX_SetupBoostIostreams.cmake
-rw-r--r-- 1 chenna chenna  835 Jul 27 11:20 HPX_SetupBoostProgramOptions.cmake
-rw-r--r-- 1 chenna chenna  608 Jul 27 11:20 HPX_SetupBoostRegex.cmake
-rw-r--r-- 1 chenna chenna 2.4K Jul 27 11:20 HPX_SetupCUDA.cmake
-rw-r--r-- 1 chenna chenna  935 Jul 27 11:20 HPX_SetupGooglePerfTools.cmake
-rw-r--r-- 1 chenna chenna 1.5K Jul 27 11:20 HPX_SetupHpxmp.cmake
-rw-r--r-- 1 chenna chenna  745 Jul 27 11:20 HPX_SetupHwloc.cmake
-rw-r--r-- 1 chenna chenna 1.5K Jul 27 11:20 HPX_SetupLibCDS.cmake
-rw-r--r-- 1 chenna chenna  11K Jul 27 11:20 HPX_SetupLibfabric.cmake
-rw-r--r-- 1 chenna chenna 1.9K Jul 27 11:20 HPX_SetupMPI.cmake
-rw-r--r-- 1 chenna chenna  813 Jul 27 11:20 HPX_SetupPapi.cmake
-rw-r--r-- 1 chenna chenna 5.8K Jul 27 11:20 HPX_SetupTarget.cmake
-rw-r--r-- 1 chenna chenna 1.1K Jul 27 11:20 HPX_SetupValgrind.cmake
-rw-r--r-- 1 chenna chenna 3.1K Jul 27 11:20 HPX_SetupVc.cmake
-rw-r--r-- 1 chenna chenna 9.2K Jul 27 11:20 HPX_SetupVerbs.cmake
-rw-r--r-- 1 chenna chenna  976 Jul 27 11:20 HPX_ShortenPseudoTarget.cmake
-rw-r--r-- 1 chenna chenna 7.1K Jul 27 11:21 HPXTargets.cmake
-rw-r--r-- 1 chenna chenna 4.7K Jul 27 11:21 HPXTargets-release.cmake
-rw-r--r-- 1 chenna chenna 5.7K Jul 27 11:20 HPX_UpdateGitDocs.cmake
-rw-r--r-- 1 chenna chenna  632 Jul 27 11:20 HPX_Utils.cmake
-rw-r--r-- 1 chenna chenna 4.8K Jul 27 11:20 installed_hpx.cmake
drwxr-xr-x 2 chenna chenna 4.0K Jul 27 11:51 scripts
-rw-r--r-- 1 chenna chenna 5.9K Jul 27 11:20 TargetArch.cmake
drwxr-xr-x 2 chenna chenna 4.0K Jul 27 11:51 tests
drwxr-xr-x 2 chenna chenna 4.0K Jul 27 11:51 toolchains
