### 2021-08-04 #3020 @diehlpk
@chennachaos Thanks for posting the output and I found the issues with the path

https://github.com/nonlocalmodels/HPCBuildInfrastructure/blob/d11ce4b6277b9b8a2a4b2030ff138e2a918cdb9d/build-nl.sh#L28

Can you please for now remove the 64 in lib64, and you should be able to compile the code. I will debug later this week why this issue with the 64 happens.
