### 2021-08-12 #3020 @diehlpk
> Hi @diehlpk. I changed the line you suggested. I also had to make a similar change at line 34 for the vtk library. After this, I got the following error. The issue is in locating the GMSH library.
>
> ```
> CMake Error at cmake/FindGmsh.cmake:19 (message):
>   Gmsh Library not found: Specify the GMSH_DIR where Gmsh is located
> Call Stack (most recent call first):
>   CMakeLists.txt:87 (find_package)
>
>
> -- Configuring incomplete, errors occurred!
> See also "/home/chenna/Documents/HPCBuildInfrastructure/src/nl/NLMech/build/CMakeFiles/CMakeOutput.log".
> See also "/home/chenna/Documents/HPCBuildInfrastructure/src/nl/NLMech/build/CMakeFiles/CMakeError.log".
> ```

@chennachaos I checked the ls of the build folder you provided

```
chenna@chenna-XPS-L501X:~/Documents/HPCBuildInfrastructure$ ls -lah build
total 48K
drwxrwxr-x 12 chenna chenna 4.0K Jul 27 13:43 .
drwxrwxr-x  6 chenna chenna 4.0K Jul 29 17:26 ..
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:03 blaze
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:04 blaze-iterative
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:08 boost
drwxrwxr-x  6 chenna chenna 4.0K Jul 27 10:54 cmake
drwxrwxr-x  7 chenna chenna 4.0K Jul 27 11:51 hpx
drwxrwxr-x  9 chenna chenna 4.0K Jul 27 11:00 hwloc
drwxr-xr-x  6 chenna chenna 4.0K Jul 27 11:02 jemalloc
drwxrwxr-x 10 chenna chenna 4.0K Jul 27 13:43 modules
drwxrwxr-x  6 chenna chenna 4.0K Jul 27 13:43 vtk
drwxrwxr-x  4 chenna chenna 4.0K Jul 27 11:14 yamlcpp
```
I can not find a `gmsh `folder there, so could it be that you did not compile gmsh using

`./build-all.sh Release without-gcc gmsh`

or installed the `gmsh-devel `package?
