### 2021-08-17 #3020 @chennachaos
@diehlpk, that's correct!
GMSH is not part of your build system. That is why I didn't install GMSH with **./build-all.sh Release without-gcc gmsh**.
There is no *build-gmsh.sh* file in the repo.
