### 2021-09-13 #3020 @chennachaos
Sorry, I didn't notice that @diehlpk already responded.

Although, all the instructions seem to be available for Fedora, not everyone uses Fedora. However, installation on different architectures is only minor issues for this library. I believe that @diehlpk will help resolve the installation issues if the users struggle to install the library on other systems.

At this point, I recommend **NLMech** library for publication in JOSS.
