### 2021-09-23 #3020 @Kevin-Mattheus-Moerman
@diehlpk I have now read your paper in detail. Below are some minor comments/suggestions on the paper. Feel free to disagree with some of these:

- [ ] Make regimes plural in `...the two regime of deformation...`
- [ ] Should `peridynamic` be written as `peridynamics`? I see several of your references do this.
- [ ] `Neighborhood of point...` -> `The neighborhood of point...`
- [ ] Question, about neighborhoods, would it be easier to understand if instead of `...is typically taken as the ball...` one would say  `...is typically defined as all points contained in the sphere...`
- [ ] `pariwise force` -> `pairwise force`
- [ ] When you say `In the library...`, I recommend referring to NLMech again since it is a while at that point since you did, just to make it clear e.g. `In the NLMech library...`
- [ ] Could this sentence be rephrased or the double example/examples be removed? `For example, regarding input files for these models, we refer to the collection of examples in the documentation.` perhaps something like: `Sample input files for these types of models are provided as examples in the documentation.` As a side note this seems to be the first mention of "input files", without having seem the documentation/software the reader is not aware that NLMech requires input files. If the examples and documentation are complete (e.g. they include a description of the use of input files) perhaps you could rephrase this further as:
`Examples for these types of model implementations are provided in the documentation.`
- [ ] `...of arbitrary domain...` this should probably read `...of an arbitrary domain...` or `...of arbitrary domains...`
- [ ] Should `...neighboring points in ball...` be rewritten as `...neighboring points in the ball...` (although I would suggest using sphere instead.
- [ ] In `...either central difference or...`, would this be better `...either a central difference or...` or `...either the central difference or...`
- [ ] Your paper is about to be accepted in JOSS, we recommend you proofread the paper once more yourself

Once you have worked through the above, please do the following
- [ ] Update the paper using `@whedon generate pdf`
- [ ] Can you let me know if the version tag (currently `v0.1.0-rc1`) should be updated?
- [ ] Archive a copy of the software on Zenodo (some find [these automated steps](https://guides.github.com/activities/citable-code/) useful) and report back here with the archive DOI.
- [ ] Can you check that the meta data like the title and author list for the Zenodo archive match the paper title/authors?
