+++ docs/paper.bib
@@ -256,3 +256,36 @@
 	year = {2020}
 }
 
+
+@article{DBLP:journals/corr/abs-2108-02336,
+  author    = {Matthias Birner and
+               Patrick Diehl and
+               Robert Lipton and
+               Marc Alexander Schweitzer},
+  title     = {A Fracture Multiscale Model for Peridynamic enrichment within the
+               Partition of Unity Method: Part {I}},
+  journal   = {CoRR},
+  volume    = {abs/2108.02336},
+  year      = {2021},
+  url       = {https://arxiv.org/abs/2108.02336},
+  eprinttype = {arXiv},
+  eprint    = {2108.02336},
+  timestamp = {Wed, 11 Aug 2021 15:24:08 +0200},
+  biburl    = {https://dblp.org/rec/journals/corr/abs-2108-02336.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@misc{bhattacharya2021peridynamics,
+      title={Peridynamics for Quasistatic Fracture Modeling}, 
+      author={Debdeep Bhattacharya and Patrick Diehl and Robert P. Lipton},
+      year={2021},
+      eprint={2107.14665},
+      archivePrefix={arXiv},
+      primaryClass={cond-mat.soft}
+}
+
+@article{diehl2021comparative,
+  title={A comparative review of peridynamics and phase-field models for engineering fracture mechanics},
+  author={Diehl, Patrick and Lipton, Robert and Wick, Thomas and Tyagi, Mayank},
+  publisher={engrXiv}
+}
