+++ docs/paper.bib
@@ -287,5 +287,6 @@
 @article{diehl2021comparative,
   title={A comparative review of peridynamics and phase-field models for engineering fracture mechanics},
   author={Diehl, Patrick and Lipton, Robert and Wick, Thomas and Tyagi, Mayank},
-  publisher={engrXiv}
+  publisher={engrXiv},
+  doi={10.31224/osf.io/gty2b}
 }
