### 2021-09-24 #3020 @diehlpk
>     * Should `peridynamic` be written as `peridynamics`? I see several of your references do this.

@Kevin-Mattheus-Moerman Yes, several references doing this, however, I decided to use peridynamic and many other authors do this.

This is written on Wikipedia for the difference between peridynamic and peridyanmic*s*.

> The term "peridynamic," an adjective, was proposed in the year 2000 and comes from the prefix peri, which means all around, near, or surrounding; and the root dyna, which means force or power. The term "peridynamics," a noun, is a shortened form of the phrase peridynamic model of solid mechanics.
