### 2021-09-27 #3020 @diehlpk
>  Your paper is about to be accepted in JOSS, we recommend you proofread the paper once more yourself

@Kevin-Mattheus-Moerman I asked some native speaker to read the paper one more time.

I will prepare the final release today and upload the paper to FigShare after my meeting marathon.
