### 2021-09-27 449091d
Add the JOSS paper (#148)
* Add template for the JOSS paper

* Add applications

* Update

* Update

* Update

* Add references

* Add application

* Backup

* Add related work

* Add material models

* Add title

* Add logo

* Add caption for logo

* Update

* Update

* Fix image

* Smaller image

* Even smaller image

* Smaller image

* Image

* Add numerical analysis reuslts

* Backup

* Add statement of need

* Add paper draft action

* Add correct path

* Add dependencies

* Add Prashant

* Test latex commands

* Add equation

* Add second equaiton

* Fix equaiton

* Fix equation

* Add e
[…]
