+++ docs/paper.bib
@@ -1,0 +1,292 @@
+@article{jha2018numerical,
+	author = {Jha, Prashant K and Lipton, Robert},
+	journal = {SIAM Journal on Numerical Analysis},
+	number = {2},
+	pages = {906--941},
+	publisher = {SIAM},
+	title = {Numerical analysis of nonlocal fracture models in Holder space},
+	volume = {56},
+	year = {2018},
+    doi={10.1137/17m1112236}
+}
+
+@article{geuzaine2009gmsh,
+	author = {Geuzaine, Christophe and Remacle, Jean-Fran{\c c}ois},
+	journal = {International journal for numerical methods in engineering},
+	number = {11},
+	pages = {1309--1331},
+	publisher = {Wiley Online Library},
+	title = {Gmsh: A 3-D finite element mesh generator with built-in pre-and post-processing facilities},
+	volume = {79},
+	year = {2009},
+    doi={10.1002/nme.2579}
+}
+
+@article{jha2019numerical,
+	author = {Jha, Prashant K and Lipton, Robert},
+	journal = {Computer Methods in Applied Mechanics and Engineering},
+	pages = {184--225},
+	publisher = {Elsevier},
+	title = {Numerical convergence of finite difference approximations for state based peridynamic fracture models},
+	volume = {351},
+	year = {2019},
+    doi={10.1016/j.cma.2019.03.024}
+}
+
+@article{lipton2019complex,
+	author = {Lipton, Robert P and Lehoucq, Richard B and Jha, Prashant K},
+	journal = {Journal of Peridynamics and Nonlocal Modeling},
+	number = {2},
+	pages = {122--130},
+	publisher = {Springer},
+	title = {Complex fracture nucleation and evolution with nonlocal elastodynamics},
+	volume = {1},
+	year = {2019},
+    doi={10.1007/s42102-019-00010-0}
+}
+
+@article{lipton2018free,
+	author = {Lipton, Robert and Said, Eyad and Jha, Prashant},
+	journal = {Journal of Elasticity},
+	number = {2},
+	pages = {129--153},
+	publisher = {Springer},
+	title = {Free damage propagation with memory},
+	volume = {133},
+	year = {2018},
+    doi={10.1007/s10659-018-9672-7}
+}
+
+@article{jha2020kinetic,
+	author = {Jha, Prashant K and Lipton, Robert P},
+	journal = {International Journal of Fracture},
+	pages = {1--15},
+	publisher = {Springer},
+	title = {Kinetic relations and local energy balance for LEFM from a nonlocal peridynamic model},
+	year = {2020},
+    doi={10.1007/s10704-020-00480-0}
+}
+
+@article{littlewood2015roadmap,
+	author = {Littlewood, David J},
+	journal = {SAND Report, Sandia National Laboratories, Albuquerque, NM and Livermore, CA},
+	title = {Roadmap for peridynamic software implementation},
+	year = {2015},
+    doi={10.2172/1226115}
+}
+
+@article{parks2008implementing,
+	author = {Parks, Michael L and Lehoucq, Richard B and Plimpton, Steven J and Silling, Stewart A},
+	journal = {Computer Physics Communications},
+	number = {11},
+	pages = {777--783},
+	publisher = {Elsevier},
+	title = {Implementing peridynamics within a molecular dynamics code},
+	volume = {179},
+	year = {2008},
+    doi={10.1016/j.cpc.2008.06.011}
+}
+
+@article{ren20173d,
+	author = {Ren, Bo and Wu, CT and Askari, E},
+	journal = {International Journal of Impact Engineering},
+	pages = {14--25},
+	publisher = {Elsevier},
+	title = {A 3D discontinuous Galerkin finite element method with the bond-based peridynamics model for dynamic brittle failure analysis},
[…]
