+++ .github/workflows/draft-pdf.yml
@@ -1,0 +1,19 @@
+on: [push]
+
+jobs:
+  paper:
+    runs-on: ubuntu-latest
+    name: Paper Draft
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+      - name: Build draft PDF
+        uses: openjournals/openjournals-draft-action@master
+        with:
+          journal: joss
+          paper-path: docs/paper.md 
+      - name: Upload
+        uses: actions/upload-artifact@v1
+        with:
+          name: paper
+          path: docs/paper.pdf
