### 2021-09-27 #3020 @diehlpk
@Kevin-Mattheus-Moerman Note, I merged the JOSS branch into master and the paper needs to be compiled from master.
