### 2021-09-29 #3020 @Kevin-Mattheus-Moerman
Congratulations @diehlpk on getting this work published in JOSS!

@vijaysm, @chennachaos thank you so much for your hard work and dedication there! This was a longer process than usual but you stuck with it and helped us throughout and did a great job providing detailed and valuable feedback so again, thank you indeed!!
